% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{AHMI}
\alias{AHMI}
\title{Adjusted Hierarchical Mutual Information}
\usage{
AHMI(tree1, tree2, Mean = max, tolerance = 0.01, minResample = 36)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}.}

\item{Mean}{Function to combine the self-information of the two trees.
Default is \code{max}.}

\item{tolerance}{Numeric specifying the tolerance for the relative error in the
expected value estimate.}

\item{minResample}{Integer specifying the minimum number of random samples to
take for the expectation calculation.}
}
\value{
A numeric value representing the adjusted hierarchical mutual information,
with an attribute \code{sem} containing the standard error of the measurement.
}
\description{
Calculate the adjusted hierarchical mutual information between two trees.
This metric adjusts the raw HMI by subtracting the expected HMI under
a random null model and normalizing by the maximum possible information.
}
\examples{
\dontrun{
library("TreeTools", quietly = TRUE)

tree1 <- BalancedTree(6)
tree2 <- PectinateTree(6)
AHMI(tree1, tree2)
}

}
\seealso{
Other tree distances: 
\code{\link{EHMI}()},
\code{\link{HierarchicalMutualInfo}()},
\code{\link{JaccardRobinsonFoulds}()},
\code{\link{KendallColijn}()},
\code{\link{MASTSize}()},
\code{\link{MatchingSplitDistance}()},
\code{\link{NNIDist}()},
\code{\link{NyeSimilarity}()},
\code{\link{PathDist}()},
\code{\link{Robinson-Foulds}},
\code{\link{SPRDist}()},
\code{\link{SelfHMI}()},
\code{\link{TreeDistance}()}
}
\concept{tree distances}
