% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\encoding{UTF-8}
\name{AllSplitPairings}
\alias{AllSplitPairings}
\title{All split pairings}
\usage{
AllSplitPairings(n)
}
\arguments{
\item{n}{Integer specifying the number of leaves in a tree.}
}
\value{
\code{AllSplitPairings} returns a named vector, specifying the number of
split pairings producing
the variation of information given (in bits) in the name.  Splits
\code{AB|CD} and \code{CD|AB} are treated as distinct, so division of all values by four
is justified in cases where unique pairings only are required.
}
\description{
Calculate the variation of (clustering) information (Meilă 2007)
for each possible pairing of non-trivial splits,
and tabulates the number of pairings with each similarity.
}
\details{
Trivial splits -- those that divide one or zero taxa from the rest --
are not considered.
}
\examples{
AllSplitPairings(6)

}
\references{
\insertRef{Meila2007}{TreeDist}

\insertRef{SmithDist}{TreeDist}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
