% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{AncestorEdge}
\alias{AncestorEdge}
\title{Ancestral edge}
\usage{
AncestorEdge(edge, parent, child)
}
\arguments{
\item{edge}{Number of an edge}

\item{parent}{parent nodes (given by phylo.object$edge[, 1])}

\item{child}{child nodes (given by phylo.object$edge[, 2])}
}
\value{
a logical vector identifying whether each edge is the edge that is ancestral to the given edge.
}
\description{
Ancestral edge
}
\keyword{internal}
