% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{ApeTime}
\alias{ApeTime}
\title{Ape Time}
\usage{
ApeTime(filename, format = "double")
}
\arguments{
\item{filename}{Character string specifying path to the file}

\item{format}{Format in which to return the time: 'double' as a sortable numeric;
any other value to return a string in the format YYYY-MM-DD hh:mm:ss}
}
\value{
The time that the specified file was created by ape.
}
\description{
Reads the time that an ape tree was modified from the comment in the nexus file
}
\author{
Martin R. Smith
}
