% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_prepare_data.R
\name{AsBinary}
\alias{AsBinary}
\title{Convert a number to binary}
\usage{
AsBinary(x)
}
\arguments{
\item{x}{Decimal integer to be converted to binary bits}
}
\value{
An array corresponding to binary digits 1, 2, 4, 8, 16, ...

'Leading zeros' are not included.
}
\description{
Provides a (reversed) binary representation of a decimal integer
}
\details{
Provides an array corresponding to binary digits 1, 2, 4, 8, 16, ...

Binary number 0100 (= decimal 4) will be represented as 0 0 1.
}
\examples{
  AsBinary(4)  # 0 0 1
  AsBinary(10) # 0 1 0 1

}
\author{
Martin R. Smith, adapted from code posted to R mailing list by Spencer Graves
}
