% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_tree_search.R
\name{BasicSearch}
\alias{BasicSearch}
\title{Search for most parsimonious trees}
\usage{
BasicSearch(tree, dataset, Rearrange = TBRCore, maxIter = 100,
  maxHits = 20, forestSize = 1, nCores = 1L, verbosity = 1, ...)
}
\arguments{
\item{tree}{a fully-resolved starting tree in \code{\link{phylo}} format, with the desired outgroup; 
edge lengths are not supported and will be deleted.}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree.}

\item{Rearrange}{Function used to rearrange trees; default: \code{\link[TreeSearch]{RootedTBR}}.}

\item{maxIter}{the maximum number of iterations to perform before abandoning the search.}

\item{maxHits}{the maximum times to hit the best score before abandoning the search.}

\item{forestSize}{the maximum number of trees to return - useful in concert with \code{\link{consensus}}.}

\item{nCores}{the number of CPU cores (processors) to employ.  As it takes a couple of seconds to set up additional
cores, you are only likely to benefit from using multiple cores when performing 
long searches on large trees.  Determine the number of available cores using 
\code{parallel::\link[parallel]{detectCores}}.}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{\dots}{other arguments to pass to subsequent functions.}
}
\value{
{
This function returns a tree, with an attribute \code{score} conveying its parsimony score.
Note that the parsimony score will be inherited from the tree's attributes, which is only valid if it 
was generated using the same \code{data} that is passed here.
}
}
\description{
Run standard search algorithms (\acronym{NNI}, \acronym{SPR} or \acronym{TBR}) 
to search for a more parsimonious tree.
}
\examples{
data('inapplicable.datasets')
my.phyDat <- inapplicable.phyData[[1]]
njtree <- TreeSearch::NJTree(my.phyDat)

\dontrun{
TreeSearch(njtree, my.phyDat, maxIter=20, Rearrange=TreeSearch::NNI)
TreeSearch(njtree, my.phyDat, maxIter=20, Rearrange=TreeSearch::RootedSPR)
}

}
\seealso{
\itemize{
\item \code{\link{InapplicableFitch}}, calculates parsimony score, supports inapplicable tokens;
\item \code{\link[TreeSearch]{RootedNNI}}, conducts tree rearrangements;
\item \code{\link{RatchetSearch}}, alternative heuristic, useful to escape local optima.
}
}
\author{
Martin R. Smith
}
\keyword{tree}
