% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{BindTree}
\alias{BindTree}
\title{Bind trees}
\usage{
BindTree(x, y, where = "root", position = 0)
}
\arguments{
\item{x}{First tree fragment}

\item{y}{Second tree fragment}

\item{where}{which edge to use to bind fragments [?]}

\item{position}{Where on second tree to bind [?]}
}
\value{
A single tree comprising the two fragments
}
\description{
Copied from \code{ape:::bind.tree}.
Changes: 
  - use (x|y).edge in place of (x|y)$edge for efficiency
  - Delete interactive option
  - Use Cladewise instead of reorder (..., 'cladewise')
}
\author{
Martin R. Smith, based on APE (Emmanuel Paradis)
}
