% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootstrapWeightings}
\alias{BootstrapWeightings}
\title{Bootstrap Weightings}
\usage{
BootstrapWeightings(attribs)
}
\arguments{
\item{attribs}{Attribtues of a phyDat object}
}
\value{
a vector giving bootstrapped weights for each character, corresponding to attribs$weights
}
\description{
Resamples edges according to their prior weights
}
\keyword{internal}
