% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brewer}
\alias{brewer}
\title{Brewer palettes}
\format{An object of class \code{list} of length 12.}
\source{
{
\href{http://colorbrewer2.org/#type=diverging&scheme=RdYlBu&n=3}{ColourBrewer2.org}
\href{http://mkweb.bcgsc.ca/colorblind/}{Martin Krzywinski}
}
}
\usage{
brewer
}
\description{
A list of eleven Brewer palettes containing one to eleven colours that
are readily distinguished by colourblind viewers, followed by a twelfth
12-colour palette adapted for colour blindness.
}
\keyword{datasets}
