% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{C_Fitch_Score}
\alias{C_Fitch_Score}
\title{Fitch Score
Calculate the parsimony score of a tree (number of steps) using the Fitch algorithm}
\usage{
C_Fitch_Score(characters, nChar, parent, child, nEdge, weight, maxNode, nTip)
}
\value{
the parsimony score (an integer)
}
\description{
Fitch Score
Calculate the parsimony score of a tree (number of steps) using the Fitch algorithm
}
\keyword{internal}
