% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{C_Fitch_Steps}
\alias{C_Fitch_Steps}
\title{Fitch steps}
\usage{
C_Fitch_Steps(characters, nChar, parent, child, nEdge, weight, maxNode, nTip)
}
\value{
the number of steps the tree enforces on each character
}
\description{
Fitch steps
}
\keyword{internal}
