% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{CalculateTreeDistance}
\alias{CalculateTreeDistance}
\title{Wrapper for tree distance calculations}
\usage{
CalculateTreeDistance(Func, tree1, tree2, reportMatching = FALSE, ...)
}
\arguments{
\item{Func}{Tree distance function.}

\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{\dots}{Additional arguments to \code{Func}.}
}
\description{
Calls tree distance functions from trees or lists of trees
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
