% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{CharH}
\alias{CharH}
\alias{CharMI}
\alias{CharJH}
\alias{CharEJH}
\alias{CharEMI}
\alias{CharAMI}
\title{Mutual information between a character and a tree}
\usage{
CharH(tree)

CharMI(char, tree)

CharJH(char, tree)

CharEJH(char, tree, precision = 0.01, minResample = 36)

CharEMI(char, tree, precision = 0.01, minResample = 36)

CharAMI(
  char,
  tree,
  Mean = function(charH, treeH) charH,
  precision = 0.01,
  minResample = 36
)
}
\arguments{
\item{tree}{Phylogenetic tree in a format that can be coerced to an \code{\link{HPart}}
object.}

\item{char}{Vector that labels each leaf in \code{tree} such that entries with
the same label share a common character state.}

\item{precision}{Numeric; Monte Carlo sampling will terminate once the
relative standard error falls below this value.}

\item{minResample}{Integer specifying minimum number of Monte Carlo samples
to conduct.  Avoids early termination when sample size is too small to
reliably estimate the standard error of the mean.}

\item{Mean}{Function by which to combine the self-information of the
two input hierarchies, in order to normalize the \acronym{HMI}.}
}
\value{
\code{CharH} returns the entropy of a tree, in bits, defined as its
capacity to assign leaves to unique clusters.
Because leaves in a cherry are not classed as distinguished, the entropy of
a tree is always less than \eqn{N \log2(N)}.

\code{CharMI} returns the mutual information content of a character and a
tree, in bits.

\code{CharJH()} returns the joint entropy of a character and a tree,
in bits, defined as the combined capacity to assign leaves to unique clusters.

\code{CharEJH()} returns the expected joint entropy of a character and a
tree, in bits, where state labels are shuffled at random.

\code{CharEMI()} returns the expected mutual information of a character
and a tree, in bits, when state labels are shuffled at random.

\code{CharEMI()} returns the expected mutual information of a character
and a tree, in bits, when state labels are shuffled at random.
}
\description{
Mutual information between a character and a tree
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
