% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{CladeSizes}
\alias{CladeSizes}
\title{Clade sizes}
\usage{
CladeSizes(tree, nodes)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{nodes}{whose descendants should be returned}
}
\value{
the number of nodes (including tips) that are descended from each node in nodes
}
\description{
Clade sizes
}
\keyword{internal}
