% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{ClusteringEntropy}
\alias{ClusteringEntropy}
\alias{ClusteringInfo}
\alias{ClusteringEntropy.phylo}
\alias{ClusteringEntropy.list}
\alias{ClusteringEntropy.multiPhylo}
\alias{ClusteringEntropy.Splits}
\alias{ClusteringInfo.phylo}
\alias{ClusteringInfo.list}
\alias{ClusteringInfo.multiPhylo}
\alias{ClusteringInfo.Splits}
\title{Clustering entropy of all splits within a tree}
\usage{
ClusteringEntropy(x)

ClusteringInfo(x)

\method{ClusteringEntropy}{phylo}(x)

\method{ClusteringEntropy}{list}(x)

\method{ClusteringEntropy}{multiPhylo}(x)

\method{ClusteringEntropy}{Splits}(x)

\method{ClusteringInfo}{phylo}(x)

\method{ClusteringInfo}{list}(x)

\method{ClusteringInfo}{multiPhylo}(x)

\method{ClusteringInfo}{Splits}(x)
}
\arguments{
\item{x}{A tree of class \code{phylo}, a list of trees, or a \code{multiPhylo} object.}
}
\value{
The sum of the entropies or (clustering) infomation content, in bits,
of each split in \code{x}.
}
\description{
Sums the entropy (\code{ClusteringEntropy()}) or information content
(\code{ClusteringInfo()}) across each split within a phylogenetic tree, treating
each split as dividing the leaves of the tree into two clusters (\emph{sensu}
Meila 2007; Vinh \emph{et al.} 2010).
}
\details{
As entropy measures the bits required to transmit the cluster label of each
leaf (Vinh \emph{et al.} 2010: p. 2840), the information content of a split is its
entropy multiplied by the number of leaves.
}
\examples{

tree1 <- TreeTools::BalancedTree(8)
tree2 <- TreeTools::PectinateTree(8)

ClusteringInfo(tree1)
ClusteringEntropy(tree1)
ClusteringInfo(list(one = tree1, two = tree2))

ClusteringInfo(tree1) + ClusteringInfo(tree2)
ClusteringEntropy(tree1) + ClusteringEntropy(tree2)
ClusteringInfoDistance(tree1, tree2)
MutualClusteringInfo(tree1, tree2)

}
\references{
\itemize{
\item \insertRef{Meila2007}{TreeDist}
\item \insertRef{Vinh2010}{TreeDist}
}
}
\seealso{
Other information functions: \code{\link{SplitEntropy}},
  \code{\link{SplitSharedInformation}},
  \code{\link{SplitwiseInfo}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{information functions}
