% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{ClusteringInfo}
\alias{ClusteringInfo}
\title{Clustering information content of all partitions within a tree}
\usage{
ClusteringInfo(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo}.}
}
\description{
Sums the clustering information content (Meila 2007) across each partition within a
phylogenetic tree.  This value will be greater than the total information
content of the tree where a tree contains multiple partitions, as
these partitions will contain mutual information.
}
\references{
\insertRef{Meila2007}{TreeSearch}
}
\author{
Martin R. Smith
}
\keyword{internal}
