% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-clustering.R
\name{ClusterMDS}
\alias{ClusterMDS}
\title{Visualize clustering structure via Multi-dimensional scaling}
\usage{
ClusterMDS(
  clust,
  distances,
  proposal = cmdscale(distances),
  maxIter = 1e+06,
  stability = 42,
  weight = list(sil = 1, mst = 1, icd = NULL),
  trace = FALSE
)
}
\arguments{
\item{maxIter}{Integer specifying maximum number of moves to propose.}

\item{stability}{After proposing \code{stability} moves without improvement,
stop.}
}
\description{
Visualize clustering structure via Multi-dimensional scaling
}
\examples{
library("TreeTools", quietly = TRUE)
treeNumbers <- c(0:20, 40:60, 120:140, 200:220)
trees <- as.phylo(treeNumbers, 8)
distances <- ClusteringInfoDist(trees)
clustering <- kmeans(distances, 4)
clust <- clustering$cluster
cmds <- cmdscale(distances)

par(mfrow = c(2, 3))
plot(cmds, col = clust, cex = 1.5, asp = 1, axes = FALSE, ann = FALSE)
plot(cluster::silhouette(clust, distances), col = 1:4, main = "Original")
plot(cluster::silhouette(clust, dist(cmds)), col = 1:4, main = "Classic MDS")

silmds <- ClusterMDS(clust, distances, cmds)
plot(silmds, col = clust, cex = 1.5, asp = 1, axes = FALSE, ann = FALSE)
MSTSegments(silmds, MSTEdges(distances),
            col = StrainCol(distances, silmds))
plot(cluster::silhouette(clust, dist(silmds)), col = 1:4, main = "SilMDS")


}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
