% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTable.R
\name{as.matrix.ClusterTable}
\alias{as.matrix.ClusterTable}
\alias{ClusterTable-methods}
\alias{print.ClusterTable}
\alias{summary.ClusterTable}
\alias{TipLabels.ClusterTable}
\alias{NTip.ClusterTable}
\alias{NSplits.ClusterTable}
\alias{SplitsInBinaryTree.ClusterTable}
\title{S3 methods for \code{ClusterTable} objects}
\usage{
\method{as.matrix}{ClusterTable}(x, ...)

\method{print}{ClusterTable}(x, ...)

\method{summary}{ClusterTable}(object, ...)

\method{TipLabels}{ClusterTable}(x, single = TRUE)

\method{NTip}{ClusterTable}(phy)

\method{NSplits}{ClusterTable}(x)

\method{SplitsInBinaryTree}{ClusterTable}(tree)
}
\arguments{
\item{x, phy, object, tree}{Object of class \code{ClusterTable}.}

\item{\dots}{Additional arguments for consistency with S3 methods.}

\item{single}{Logical specifying that labels should be returned from the
first (and only) entry in \code{x}.}
}
\description{
S3 methods for \code{ClusterTable} objects
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
