% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-clustering.R
\name{ClusVis}
\alias{ClusVis}
\title{Visualize clustering distances}
\usage{
ClusVis(distances, clust, col = seq_len)
}
\arguments{
\item{distances}{Matrix or \code{dist} object giving distance between each
pair of trees.}

\item{clust}{Integer vector listing the sequentially numbered cluster to
which each tree is assigned.}

\item{col}{Colours with which to colour each cluster.
Either specify a vector of colours with an entry for each cluster,
or a function that will generate a palette when passed a variable denoting
the number of colours required.}
}
\description{
Visualize clustering distances
}
\examples{
library("TreeTools")
treeNumbers <- c(0:20, 40:60, 120:140, 200:220)
group <- rep(1:4, each = 21)
trees <- as.phylo(treeNumbers, 8)
distances <- ClusteringInfoDist(trees)
set.seed(1) # Bad clustering
clustering <- kmeans(distances, 4)
plot(cmdscale(distances), col = clustering$cluster, cex = 1.5,
     asp = 1, axes = FALSE, ann = FALSE)
ClusVis(distances, clustering$cluster)
plot(cluster::silhouette(clustering$cluster, distances),
     col = 1:4)

set.seed(2) # Good clustering
clustering <- kmeans(distances, 4)
Cols <- function (n) hcl.colors(n, pal = "Dynamic")
plot(cmdscale(distances), col = Cols(4)[clustering$cluster], cex = 1.5,
     asp = 1, axes = FALSE, ann = FALSE)

ClusVis(distances, clustering$cluster, col = Cols)
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
