% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{CollapseSingles}
\alias{CollapseSingles}
\title{Collapse Single Nodes}
\usage{
CollapseSingles(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}
}
\description{
\code{CollapseSingles} is duplicated from ape::collapse.singles and modified to improve
its running speed.
}
\author{
Martin R. Smith
}
