% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congreveLamsdellMatrices}
\alias{congreveLamsdellMatrices}
\title{100 simulated data matrices}
\format{A list with 100 entries, each comprising a phyDat object of 55 characters for 22 taxa}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.7dq0j}
}
\usage{
congreveLamsdellMatrices
}
\description{
Contains the 100 simulated matrices generated by Congreve & Lamsdell (2016) using a
heterogeneous Markov-k model, generated from the \link{referenceTree} topology, with
all branches sharing an equal length.
}
\examples{
  data('referenceTree')
  data('congreveLamsdellMatrices')
  \dontrun{ProfileScore(referenceTree,
   PrepareDataProfile(congreveLamsdellMatrices[[17]]))}

}
\references{
\insertRef{Congreve2016}{TreeSearch}
\insertRef{Congreve2016dd}{TreeSearch}
}
\keyword{datasets}
