% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{DescendantEdges}
\alias{DescendantEdges}
\title{Descendant Edges}
\usage{
DescendantEdges(edge, parent, child, nEdge = length(parent))
}
\arguments{
\item{edge}{number of the edge whose child edges are required}

\item{parent}{the first column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 1]}

\item{child}{the second column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 2]}

\item{nEdge}{number of edges (calcluated from length(parent) if not supplied)}
}
\value{
a logical vector stating whether each edge in turn is a descendant of the speficied edge
        (or the edge itself)
}
\description{
Quickly identifies edges that are 'descended' from a particular edge in a tree
}
