% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distanceDistribution25}
\alias{distanceDistribution25}
\title{Distances between random pairs of 25-tip trees}
\format{An object of class \code{matrix} with 9 rows and 10000 columns.}
\usage{
distanceDistribution25
}
\description{
A two-dimensional matrix listing the normalized distances between 10000 random
pairs of 25-tip trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br=NULL)}.
}
\details{
Rows are named with an abbreviation of the tree comparison metric.TODO list these and state how normalized.

Columns list the calculated tree distances.
}
\keyword{datasets}
