% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{DoDescendants}
\alias{DoDescendants}
\title{TITLE GOES HERE}
\usage{
DoDescendants(edge1, edge2, nTip, node, just.tips = FALSE,
  just.internal = FALSE, include.ancestor = FALSE)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{FUNCTIONNAME} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
