% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{.CalculateHMIRecursive}
\alias{.CalculateHMIRecursive}
\title{Calculate Hierarchical Mutual Information recursively}
\usage{
.CalculateHMIRecursive(Ut, Us)
}
\arguments{
\item{Ut, Us}{Hierarchical partitions (nested lists)}
}
\value{
List with n_ts and I_ts values
}
\description{
Calculate Hierarchical Mutual Information recursively
}
\keyword{internal}
