% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{.HMI11}
\alias{.HMI11}
\title{Hierarchical Mutual Information between splits}
\usage{
.HMI11(tree1, tree2, normalize)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, with leaves labelled identically,
or lists of such trees to undergo pairwise comparison.  Where implemented,
\code{tree2 = NULL} will compute distances between each pair of trees in the list
\code{tree1} using a fast algorithm based on
\insertCite{Day1985;textual}{TreeDist}.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the "Normalization"
section below.
If \code{FALSE}, results will not be rescaled.}

\item{splits1, splits2}{Logical matrices where each row corresponds to a split
on the tree, as produced by \code{\link[TreeTools:Splits]{TreeTools::as.Splits()}}.}

\item{nTip}{Integer specifying the number of tips in the trees.}

\item{reportMatching}{Logical specifying whether to report the matching
between splits.}
}
\value{
A numeric value representing the Hierarchical Mutual Information
between the trees.
}
\description{
Calculate the Hierarchical Mutual Information between splits representations
of two trees.
}
\keyword{internal}
