% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_mast.R
\name{.MASTSizeEdges}
\alias{.MASTSizeEdges}
\title{Caluclate MAST size from edge matrices.}
\usage{
.MASTSizeEdges(edge1, edge2, nTip)
}
\arguments{
\item{edge1}{Edge matrix of tree 1. MUST BE IN POSTORDER!}

\item{edge2}{Edge matrix of tree 2.}

\item{nTip}{Integer specifying the number of leaves in each split.}
}
\description{
Internal function.
}
\keyword{internal}
