% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{.PhyloToHierarchicalPartition}
\alias{.PhyloToHierarchicalPartition}
\title{Convert phylo tree to hierarchical partition}
\usage{
.PhyloToHierarchicalPartition(tree)
}
\arguments{
\item{tree}{A phylo object}
}
\value{
A nested list representing the hierarchical partition
}
\description{
Convert phylo tree to hierarchical partition
}
\keyword{internal}
