% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doubleFactorials}
\alias{doubleFactorials}
\title{Double factorials}
\format{An object of class \code{numeric} of length 300.}
\usage{
doubleFactorials
}
\description{
A vector with pre-calculated values of double factorials up to 300!!,
and the logarithms of double factorials up to 50 000!!.
}
\details{
301!! is too large to store as an integer; use \code{logDoubleFactorials} instead.
}
\seealso{
Other Double factorial: \code{\link{DoubleFactorial}},
  \code{\link{logDoubleFactorials}}
}
\concept{Double factorial}
\keyword{datasets}
