% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNI.R
\name{DoubleNNI}
\alias{DoubleNNI}
\title{Double NNI}
\usage{
DoubleNNI(parent, child, edgeToBreak)
}
\arguments{
\item{parent}{the first column of the edge matrix of a tree of class
\code{\link{phylo}}, i.e. \code{tree$edge[, 1]}}

\item{child}{the second column of the edge matrix of a tree of class
\code{\link{phylo}}, i.e. \code{tree$edge[, 2]}}

\item{edgeToBreak}{(optional) integer specifying the index of an edge to bisect/prune,
generated randomly if not specified.
Alternatively, set to \code{-1} to return a complete list
of all trees one step from the input tree.}
}
\value{
the \code{tree$edge} parameter of the two trees consistent with the specified rearrangement
}
\description{
Returns the edge parameter of the two trees consistent with the speficied \acronym{NNI} rearrangement
}
\author{
Martin R. Smith
}
\keyword{internal}
