% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{DropTipNoSubtree}
\alias{DropTipNoSubtree}
\title{TITLE GOES HERE}
\usage{
DropTipNoSubtree(phy, tip, root.edge = 0, rooted = is.rooted(phy),
  interactive = FALSE)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{FUNCTIONNAME} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
