% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{EdgeAncestry}
\alias{EdgeAncestry}
\title{Descendant Edges}
\usage{
EdgeAncestry(edge, parent, child, stopAt = (parent == min(parent)))
}
\arguments{
\item{edge}{number of the edge whose child edges are required}

\item{parent}{the first column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 1]}

\item{child}{the second column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 2]}

\item{stopAt}{number of the edge at which the search should terminate; defaults to the root edges}

\item{nEdge}{number of edges (calcluated from length(parent) if not supplied)}
}
\value{
a logical vector stating whether each edge in turn is a descendant of the speficied edge
}
\description{
Quickly identifies edges that are 'ancestral' to a particular edge in a tree
}
