% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{EHMI}
\alias{EHMI}
\title{Expected Hierarchical Mutual Information}
\usage{
EHMI(tree1, tree2, tolerance = 0.01, minResample = 36)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}.}

\item{tolerance}{Numeric specifying the tolerance for the relative error in the
expected value estimate.}

\item{minResample}{Integer specifying the minimum number of random samples to
take for the expectation calculation.}
}
\value{
A numeric vector with the expected hierarchical mutual information,
with attributes containing variance, standard deviation, standard error,
number of samples, and relative error.
}
\description{
Calculate the expected hierarchical mutual information between two trees
under a null model where tip labels are randomly shuffled.
}
\examples{
\dontrun{
library("TreeTools", quietly = TRUE)

tree1 <- BalancedTree(6)
tree2 <- PectinateTree(6)
EHMI(tree1, tree2)
}

}
\seealso{
Other tree distances: 
\code{\link{AHMI}()},
\code{\link{HierarchicalMutualInfo}()},
\code{\link{JaccardRobinsonFoulds}()},
\code{\link{KendallColijn}()},
\code{\link{MASTSize}()},
\code{\link{MatchingSplitDistance}()},
\code{\link{NNIDist}()},
\code{\link{NyeSimilarity}()},
\code{\link{PathDist}()},
\code{\link{Robinson-Foulds}},
\code{\link{SPRDist}()},
\code{\link{SelfHMI}()},
\code{\link{TreeDistance}()}
}
\concept{tree distances}
