% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{EnforceOutgroup}
\alias{EnforceOutgroup}
\title{Force taxa to form an outgroup}
\usage{
EnforceOutgroup(tree, outgroup)
}
\arguments{
\item{tree}{either a tree of class \code{phylo}, or a character vector listing the names of 
all the taxa in the tree, from which a random tree will be generated.}

\item{outgroup}{a vector containing the names of taxa to include in the outgroup}
}
\value{
a tree where all outgroup taxa are sister to all remaining taxa, 
        otherwise retaining the topology of the ingroup.
}
\description{
Given a tree or a list of taxa, rearrange the ingroup and outgroup taxa such that the two
are sister taxa across the root, without altering the relationships within the ingroup
or within the outgroup.
}
\author{
Martin R. Smith
}
