% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{ExtractClade}
\alias{ExtractClade}
\title{ExtractClade}
\usage{
ExtractClade(phy, node)
}
\arguments{
\item{phy}{A phylogenetic tree in \code{phylo} format;}

\item{node}{The number of the node at the base of the clade to be extracted.}
}
\value{
This function returns a phylogenetic tree that represents a clade extracted from the original tree.
}
\description{
Safely extracts a clade from a phylogenetic tree.
}
\details{
Extract a clade


Modified from the \pkg{ape} function \code{\link{extract.clade}}, which sometimes behaves erratically.  
The function is intended for use with the function \code{\link{TBR}}, and features present in 
\code{extract.clade} but unnecessary for this goal have been removed. Unlike extract.clade, 
this function supports the extraction of 'clades' that constitute a single tip.
}
\examples{
{
library('phangorn')
tree <- rtree(20, br=NULL)
plot(tree); nodelabels(); nodelabels(33, 33, bg='yellow'); dev.new()
plot(ExtractClade(tree, 33))
}

}
\seealso{
extract.clade
}
\author{
Martin R. Smith
}
