% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{ExtractTaxa}
\alias{ExtractTaxa}
\title{Extract taxa from a matrix block}
\usage{
ExtractTaxa(matrixLines, character_num = NULL, session = NULL)
}
\arguments{
\item{matrixLines}{lines of a file containing a phylogenetic matrix
(see ReadCharacters for expected format)}

\item{character_num}{Index of character(s) to return.
`NULL`, the default, returns all characters.}

\item{session}{(optionally) a Shiny session with a numericInput named character_num
whose maximum should be updated.}
}
\value{
Matrix with n rows, each named for the relevant taxon, and c columns,
each corresponding to the respective character specified in `character_num`
}
\description{
Extract taxa from a matrix block
}
\keyword{internal}
