% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{Fitch}
\alias{Fitch}
\alias{FitchScore}
\alias{FasterFitchScore}
\alias{FitchSteps}
\title{Fitch score}
\usage{
Fitch(tree, data, TipData = TipsAreNames, at = attributes(data),
  FitchFunction = C_Fitch_Score)

FitchScore(tree, data, TipData = NULL, at = attributes(data))

FasterFitchScore(tree, data, TipData = TipsAreNames, at = attributes(data))

FitchSteps(tree, data, TipData = TipsAreNames, at = attributes(data))
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{data}{A list or matrix whose list entries, rows or columns (as specified in TipData)
correspond to the character data associated with the tips of the tree. 
Likely of class \code{phyDat}.}

\item{TipData}{function to be used to match tree tips to dataset; probably one of 
\code{TipsAreNames}, \code{TipsAreRows}, or \code{TipsAreColumns}}

\item{at}{Attributes of the dataset (looked up automatically if not supplied)}

\item{FitchFunction}{function to be used to calculte parsimony score.}
}
\value{
A vector listing the number of 'parsimony steps' calculated for each character
}
\description{
Fitch score
}
\section{Functions}{
\itemize{
\item \code{FitchScore}: returns the parsimony score only

\item \code{FasterFitchScore}: returns the parsimony score only, without checking that data is well formatted

\item \code{FitchSteps}: returns a vector listing the number of steps for each character
}}

\keyword{internal}
