% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{GetAncestors}
\alias{GetAncestors}
\title{Get Ancestors}
\usage{
GetAncestors(parent, child, node)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{GetAncestors} gets the ancestors of each node in a tree
It's a more efficient version of \code{\link[phangorn]{Ancestors}}
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
