% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{GetDescendants}
\alias{GetDescendants}
\title{Get Descendants}
\usage{
GetDescendants(tree, node, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{node}{the number of the internal node whose descendants should be returned}
}
\value{
This function returnsa vector containing descendant nodes in numerical order
}
\description{
\code{GetDescendants} returns the descendants of a specified node
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
