% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H_xptr}
\alias{H_xptr}
\alias{JH_xptr}
\alias{EJH_xptr}
\alias{EMI_xptr}
\alias{AMI_xptr}
\title{Directly calculate entropy from an HPart pointer}
\usage{
H_xptr(ptr)

JH_xptr(char_ptr, tree_ptr)

EJH_xptr(char_ptr, tree_ptr, precision = 0.01, minResample = 36L)

EMI_xptr(char_ptr, tree_ptr, precision = 0.01, minResample = 36L)

AMI_xptr(char_ptr, tree_ptr, mean_fn, precision = 0.01, minResample = 36L)
}
\arguments{
\item{ptr, char_ptr, tree_ptr}{Pointers to HPart objects for which chosen
entropy value should be calculated.}

\item{precision}{Numeric; Monte Carlo sampling will terminate once the
relative (or for \code{CharAMI()}, absolute) standard error falls below this
value.}

\item{minResample}{Integer specifying minimum number of Monte Carlo samples
to conduct.  Avoids early termination when sample size is too small to
reliably estimate the standard error of the mean.}
}
\description{
Intended for developer use only; no safeguards against bad input.
May crash R if used incorrectly.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
