% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_hvi.R
\name{HierachicalVariation}
\alias{HierachicalVariation}
\title{Hierarchical Variation of Information distance}
\usage{
HierachicalVariation(tree1, tree2 = NULL)
}
\description{
Calculate the hierachicical variation of information distance
}
\details{
Explain here how the hierachical variation of information distance works
}
