% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{HierarchicalMutualInformation}
\alias{HierarchicalMutualInformation}
\title{Hierarchical mutual information}
\usage{
HierarchicalMutualInformation(tree1, tree2 = NULL, normalize = FALSE)
}
\description{
An alias of \code{HierarchicalMutualInfo()}
}
\keyword{internal}
