% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{HierarchicalMutualInfoSplits}
\alias{HierarchicalMutualInfoSplits}
\title{Calculate Hierarchical Mutual Information between splits}
\usage{
HierarchicalMutualInfoSplits(
  splits1,
  splits2,
  nTip = attr(splits1, "nTip"),
  normalize = FALSE,
  reportMatching = FALSE
)
}
\arguments{
\item{splits1, splits2}{Objects of class \code{Splits}.}

\item{nTip}{Integer specifying the number of tips.}

\item{normalize}{Logical. If \code{TRUE}, normalize the result.}

\item{reportMatching}{Logical specifying whether to return matchings.}
}
\value{
Numeric value of Hierarchical Mutual Information.
}
\description{
Calculate Hierarchical Mutual Information between splits
}
