% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{HMI}
\alias{HMI}
\title{Computes the hierarchical mutual information between two hierarchical partitions.}
\usage{
HMI(Ut, Us)
}
\value{
Returns
n_ts,HMI(Ut,Us) : where n_ts is the number of common elements between the
hierarchical partitions Ut and Us.

NOTE: We label by u,v the children of t,s respectively.
}
\description{
Computes the hierarchical mutual information between two hierarchical partitions.
}
