% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{ICS}
\alias{ICS}
\alias{ICPerStep}
\title{Information content per step}
\usage{
ICS(..., envir = parent.frame())

ICPerStep(splits, maxIter, warn = TRUE)
}
\arguments{
\item{splits}{A vector of integers listing the number of tips in each of a number of
tree splits (e.g. bipartitions).  For example, \kbd{c(3, 5)} states that
a character divides a set of eight tips into a group of three and a group
of five.}

\item{maxIter}{number of iterations to use when estimating concavity constant}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity
functions are generated by approximation.}

\item{a}{\code{min(splits)}.}

\item{b}{\code{max(splits)}.}

\item{m}{\code{maxIter}.}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity
functions are generated by approximation.}
}
\description{
Information content per step
}
\section{Functions}{
\itemize{
\item \code{ICS}: Memoized calculating function
}}

\keyword{internal}
