% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inapplicable.datasets}
\alias{inapplicable.datasets}
\title{Thirty Datasets with Inapplicable data}
\format{An object of class \code{list} of length 30.}
\source{
{
\describe{
\item{Agnarsson2004}{AGNARSSON, I. 2004. Morphological phylogeny of cobweb spiders and their
relatives (Araneae, Araneoidea, Theridiidae). Zoological Journal of the
Linnean Society, 141, 447--626.}
\item{Capa2011}{CAPA, M., HUTCHINGS, P., AGUADO, M. T. and BOTT, N. J. 2011. Phylogeny of
Sabellidae (Annelida) and relationships with other taxa inferred from
morphology and multiple genes. Cladistics, 27, 449--469.}
\item{DeAssis2011}{ DE ASSIS, J. E. and CHRISTOFFERSEN, M. L. 2011. Phylogenetic relationships
within Maldanidae (Capitellida, Annelida), based on morphological characters.
Systematics and Biodiversity, 9, 233--245.}
\item{OLeary1999}{ O'LEARY, M. A. and GEISLER, J. H. 1999. The position of Cetacea within
Mammalia: phylogenetic analysis of morphological data from extinct and
extant taxa. Systematic Biology, 48, 455--490.}
\item{Rousset2004}{ ROUSSET, V., ROUSE, G. W., SIDDALL, M. E., TILLIER, A. and PLEIJEL, F.
2004. The phylogenetic position of Siboglinidae (Annelida) inferred from
18S rRNA, 28S rRNA and morphological data. Cladistics, 20, 518--533.}
\item{Sano2011}{ SANO, M. and AKIMOTO, S.-I. 2011. Morphological phylogeny of
gall-forming aphids of the tribe Eriosomatini (Aphididae: Eriosomatinae).
Systematic Entomology, 36, 607--627.}
\item{Sansom2010}{ SANSOM, R. S., FREEDMAN, K., GABBOTT, S. E., ALDRIDGE, R. J. and
PURNELL, M. A. 2010. Taphonomy and affinity of an enigmatic Silurian vertebrate,
\emph{Jamoytius kerwoodi} White. Palaeontology, 53, 1393--1409.}
\item{Schulze2007}{ SCHULZE, A., CUTLER, E. B. and GIRIBET, G. 2007. Phylogeny of sipunculan
worms: A combined analysis of four gene regions and morphology. Molecular Phylogenetics and Evolution, 42, 171--92.}
\item{Shultz2007}{ SHULTZ, J. W. 2007. A phylogenetic analysis of the arachnid orders based
on morphological characters. Zoological Journal of the Linnean Society, 150, 221--265.}
\item{Wetterer2000}{ WETTERER, A. L., ROCKKMAN, M. V. and SIMMONS, N. B. 2000. Phylogeny of
phyllostomid bats (Mammalia: Chiroptera): data from diverse morphological
systems, sex chromosomes, and restriction sites. Bulletin of the American Museum of Natural History, 248, 1--200.}
\item{Wills2012}{ WILLS, M. A., GERBER, S., RUTA, M. and HUGHES, M. 2012. The disparity of
priapulid, archaeopriapulid and palaeoscolecid worms in the light of new data.
Journal of Evolutionary Biology, 25, 2056--2076.}
\item{Aguado2009}{ AGUADO, M. T. and SAN MARTIN, G. 2009. Phylogeny of Syllidae (Polychaeta)
based on morphological data. Zoologica Scripta, 38, 379--402.}
\item{Aria2015}{ ARIA, C., CARON, J. B. and GAINES, R. 2015. A large new leanchoiliid from
the Burgess Shale and the influence of inapplicable states on stem arthropod
phylogeny. Palaeontology, 58, 629--660.}
\item{Asher2005}{ASHER, R. J. and HOFREITER, M. 2006. Tenrec phylogeny and the noninvasive
extraction of nuclear DNA. Systematic biology, 55, 181--94.}
\item{Baker2009}{ BAKER, W. J., SAVOLAINEN, V., ASMUSSEN-LANGE, C. B., CHASE, M. W.,
DRANSFIELD, J., FOREST, F., HARLEY, M. M., UHL, N. W. and WILKINSON, M. 2009.
Complete generic-level phylogenetic analyses of palms (Arecaceae) with
comparisons of supertree and supermatrix approaches.
Systematic Biology, 58, 240--256.}
\item{Bouchenak2010}{ BOUCHENAK-KHELLADI, Y., VERBOOM, G. A., SAVOLAINEN, V. and HODKINSON, T. R.
2010. Biogeography of the grasses (Poaceae): a phylogenetic approach to reveal
evolutionary history in geographical space and geological time. Botanical
Journal of the Linnean Society, 162, 543--557.}
\item{Conrad2008}{ CONRAD, J. L. 2008. Phylogeny And Systematics Of Squamata (Reptilia) Based
On Morphology. Bulletin of the American Museum of Natural History, 310, 1--182.}
\item{Dikow2009}{ DIKOW, T. 2009. A phylogenetic hypothesis for Asilidae based on a total
evidence analysis of morphological and DNA sequence data (Insecta: Diptera:
Brachycera: Asiloidea). Organisms Diversity and Evolution, 9, 165--188.}
\item{Eklund2004}{ EKLUND, H., DOYLE, J. A. and HERENDEEN, P. S. 2004. Morphological
phylogenetic analysis of living and fossil Chloranthaceae. International
Journal of Plant Sciences, 165, 107--151.}
\item{Geisler2001}{ GEISLER, J. H. 2001. New morphological evidence for the phylogeny of
Artiodactyla, Cetacea, and Mesonychidae. American Museum Novitates, 3344, 53.}
\item{Giles2015}{ GILES, S., FRIEDMAN, M. and BRAZEAU, M. D. 2015. Osteichthyan-like cranial
conditions in an Early Devonian stem gnathostome. Nature, 520, 82--85.}
\item{Griswold1999}{ GRISWOLD, C. E., CODDINGTON, J. A., PLATNICK, N. I. and FORSTER, R. R. 1999.
Towards a phylogeny of entelegyne spiders (Araneae, Araneomorphae, Entelegynae).
Journal of Arachnology, 27, 53--63.}
\item{Liljeblad2008}{ LILJEBLAD, J., RONQUIST, F., NIEVES-ALDREY, J. L., FONTAL-CAZALLA, F.,
ROS-FARRE, P., GAITROS, D. and PUJADE-VILLAR, J. 2008. A fully web-illustrated
morphological phylogenetic study of relationships among oak gall wasps and
their closest relatives (Hymenoptera: Cynipidae).}
\item{Loconte1991}{ LOCONTE, H. and STEVENSON, D. W. 1991. Cladistics of the Magnoliidae.
Cladistics, 7, 267--296.}
\item{Longrich2010}{ LONGRICH, N. R., SANKEY, J. and TANKE, D. 2010. \emph{Texacephale langstoni},
a new genus of pachycephalosaurid (Dinosauria: Ornithischia) from the upper
Campanian Aguja Formation, southern Texas, USA. Cretaceous Research, 31, 274--284.}
\item{OMeara2014}{ O'MEARA, R. N. and THOMPSON, R. S. 2014. Were There Miocene Meridiolestidans?
Assessing the phylogenetic placement of \emph{Necrolestes patagonensis} and the
presence of a 40 million year Meridiolestidan ghost lineage. Journal
of Mammalian Evolution, 21, 271--284.}
\item{Rougier2012}{ ROUGIER, G. W., WIBLE, J. R., BECK, R. M. D. and APESTEGUIA, S. 2012.
The Miocene mammal \emph{Necrolestes} demonstrates the survival of a
Mesozoic nontherian lineage into the late Cenozoic of South America.
Proceedings of the National Academy of Sciences, 109, 20053--8.}
\item{Sharkey2011}{ SHARKEY, M. J., CARPENTER, J. M., VILHELMSEN, L., HERATY, J., LILJEBLAD, J.,
DOWLING, A. P. G., SCHULMEISTER, S., MURRAY, D., DEANS, A. R., RONQUIST, F.,
KROGMANN, L. and WHEELER, W. C. 2012. Phylogenetic relationships among
superfamilies of Hymenoptera. Cladistics, 28, 80--112.}
\item{Sundue2010}{ SUNDUE, M. A., ISLAM, M. B. and RANKER, T. A. 2010. Systematics of
Grammitid Ferns (Polypodiaceae): Using Morphology and Plastid Sequence Data
to Resolve the Circumscriptions of Melpomene and the Polyphyletic Genera
\emph{Lellingeria} and \emph{Terpsichore}. Systematic Botany, 35, 701--715.}
\item{Vinther2008}{ VINTHER, J., VAN ROY, P. and BRIGGS, D. E. G. 2008. Machaeridians are
Palaeozoic armoured annelids. Nature, 451, 185--188.}
\item{Wilson2003}{ WILSON, G. D. F. and EDGECOMBE, G. D. 2003. The Triassic isopod
\emph{Protamphisopus wianamattensis} (Chilton) and comparison by extant
taxa (Crustacea, Phreatoicidea). Journal of Paleontology, 77, 454--470.}
\item{Wortley2006}{ WORTLEY, A. H. and SCOTLAND, R. W. 2006. The effect of combining molecular
and morphological data in published phylogenetic analyses.
Systematic Biology, 55, 677--685.}
\item{Zanol2014}{ ZANOL, J., HALANYCH, K. M. and FAUCHALD, K. 2014. Reconciling taxonomy and
phylogeny in the bristleworm family Eunicidae (Polychaete, Annelida).
Zoologica Scripta, 43, 79--100.}
\item{Zhu2013}{ ZHU, M., YU, X., AHLBERG, P. E., CHOO, B., LU, J., QIAO, T., QU, Q., ZHAO, W.,
JIA, L., BLOM, H. and ZHU, Y. 2013. A Silurian placoderm with osteichthyan-like
marginal jaw bones. Nature, 502, 188--193.}
}
}
}
\usage{
inapplicable.datasets
}
\description{
These are the datasets used to evaluate the behaviour of the inapplicable algorithm
in Brazeau, Guillerme and Smith (2017).
}
\details{
The name of each item corresponds to the datasets listed below.
Datasets are sorted into two subsets, each sorted alphabetically;
the first subset comprise simpler datasets with faster processing times.
The value is the dataset in the format generated by \code{\link{read.nexus.data}}.
}
\references{
\insertRef{Brazeau2018}{TreeSearch}
}
\keyword{datasets}
