% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{InfoAmounts}
\alias{InfoAmounts}
\title{Amount of information in each character}
\usage{
InfoAmounts(tokenTable, precision = 1e+06, warn = TRUE)
}
\arguments{
\item{tokenTable}{A matrix, where each row corresponds to a character, each column to
a tip, and each entry to the value (1 or 2) of the character at that tip.}

\item{precision}{number of random trees to generate when calculating Profile curves}

\item{warn}{Boolean (\code{TRUE}/\code{FALSE}): display warnings when concavity
functions are generated by approximation.}
}
\value{
information content of each extra step, in bits
}
\description{
As presently implemented, this function requires that there be no ambiguous tokens
and two applicable tokens, '1' and '2'.
}
\author{
Martin R. Smith
}
