% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{KeepEdges}
\alias{KeepEdges}
\title{TITLE GOES HERE}
\usage{
KeepEdges(edge, tip.label, nTips, kept.edges)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{FUNCTIONNAME} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
