% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{LDFact}
\alias{LDFact}
\title{Log double factorial
Handles odd and even inout numbers}
\usage{
LDFact(x)
}
\arguments{
\item{x}{a positive integer}
}
\description{
Log double factorial
Handles odd and even inout numbers
}
