% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{LDFactorial}
\alias{LDFactorial}
\title{Log double factorial}
\usage{
LDFactorial(x)
}
\arguments{
\item{x}{(integer) number to crunch.}
}
\description{
Memoised version of phangorn's \code{\link[phangorn]{ldfactorial}}
}
