% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ListToMatrix}
\alias{ListToMatrix}
\title{Edge list to edge matrix
Converts list(parent, child) to the format expected by tree$edge.}
\usage{
ListToMatrix(edgeList, nEdge = NULL)
}
\description{
Edge list to edge matrix
Converts list(parent, child) to the format expected by tree$edge.
}
