% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{LoadMorphy}
\alias{LoadMorphy}
\title{Initialize a Morphy Object from a phyDat object}
\usage{
LoadMorphy(phy)
}
\arguments{
\item{phy}{an object of class \code{\link{phyDat}}}
}
\value{
a pointer to a Morphy object, with the attribute "weight" corres
}
\description{
Creates a new Morphy object with the same size and characters as the phyDat object
}
\author{
Martin R. Smith
}
