% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lobo.data}
\alias{Lobo.data}
\title{Raw data from Zhang et al. 2016}
\format{An object of class \code{list} of length 48.}
\source{
\insertRef{Zhang2016}{TreeSearch}
}
\usage{
Lobo.data
}
\description{
Raw data from Zhang et al. 2016

Modified so that absences are treated appropriately:
\describe{
\item{Character 7 }{inapplicable to absent where cephalic shield (char 3) is absent}
\item{Character 40}{inapplicable to absent where paired appendages absent}
\item{Character 46}{inapplicable to absent}
\item{Character 64}{76: stet; trunk annulations / limbs may primitively have been papillate or non-papillate}
\item{Character 69}{a good case that the fusion of flaps with endopods is secondary; thus inapplicable to absent where 67 is applicable}
\item{Character 72}{inapplicable to absent where appendages are present}
\item{Character 77}{inapplicable to absent where papillae applicable, as spine is a secondary elaboration of papillae}
\item{Character 78}{inapplicable to absent}
\item{Character 79}{stet, as possible that limbs evolved by extension of plate-like exoskeletal element }
\item{Character 80}{inapplicable to absent, on assumption that ancestral claws were simple }
\item{Characters 83, 84, 86, 87, 92}{inapplicable to absent; an obvious elaboration }
\item{Character 93}{inapplicable to absent; ancestrally undifferentiated by definition? }
\item{Character 96}{inapplicable to absent; ancestrally undifferentiated by definition. Ambiguous in taxa that lack claws as rotation would not be observed. ]}
}
}
\keyword{datasets}
