% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logDoubleFactorials}
\alias{logDoubleFactorials}
\title{Natural logarithms of double factorials}
\format{An object of class \code{numeric} of length 50000.}
\usage{
logDoubleFactorials
}
\description{
A vector with pre-calculated values of double factorials up to 50 000!!.
}
\seealso{
Other Double factorial: \code{\link{DoubleFactorial}},
  \code{\link{doubleFactorials}}
}
\concept{Double factorial}
\keyword{datasets}
