% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{LogisticPoints}
\alias{LogisticPoints}
\title{Logistic Points
Extract points from a fitted model}
\usage{
LogisticPoints(x, fittedModel)
}
\arguments{
\item{x}{an integer vector giving x co-ordinates.}

\item{fittedModel}{a fitted model, perhaps generated using \kbd{nls(cumP ~ SSlogis(nSteps, Asym, xmid, scal))}.}
}
\value{
values of y co-ordinates corresponding to the x co-ordinates provided
}
\description{
Logistic Points
Extract points from a fitted model
}
\author{
Martin R. Smith
}
