% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_mast.R
\name{MAST}
\alias{MAST}
\title{Size of maximum agreement subtree}
\usage{
MAST(tree1, tree2, tree = FALSE, rooted = TRUE)
}
\arguments{
\item{tree1, tree2}{Single trees of class \code{phylo} to undergo comparison.}
}
\description{
Calculate the maximum agreement subtree between two phylogenetic trees.
}
\references{
\insertref{Valiente2009}{TreeDist}
}
\seealso{
\code{\link[phangorn:mast]{phangorn::mast}}, a slower, all-R implementation
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
