% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{MatrixToList}
\alias{MatrixToList}
\title{Edge matrix to edge list}
\usage{
MatrixToList(edge)
}
\arguments{
\item{edge}{edges in the matrix format used by \code{tree$edge}, where \code{tree} is a tree of class \code{phylo}}
}
\value{
tree edges in the format list(parent, child).
}
\description{
Edge matrix to edge list
}
