% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{MatrixToPhyDat}
\alias{MatrixToPhyDat}
\title{Matrix to phydat}
\usage{
MatrixToPhyDat(tokens)
}
\arguments{
\item{tokens}{matrix of tokens, probably created with \link{ReadCharacters}
or \link{ReadTntCharacters}.}
}
\value{
an object of class \code{phyDat}
}
\description{
Converts a matrix of tokens to a phyDat object
}
\author{
Martin R. Smith
}
\keyword{internal}
