% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_data_manipulation.R
\name{MinimumSteps}
\alias{MinimumSteps}
\title{Minimum steps}
\usage{
MinimumSteps(states)
}
\arguments{
\item{states}{Integer vector listing the tokens that may be present at each tip along a single character,
with each token represented as a binary digit; e.g. a value of 11 means that
the tip may have tokens 0, 1 or 3 (as 11 = 2^0 + 2^1 + 2^3).
As the minimum steps can be found when inapplicables occur together,
inapplicable tokens can be denoted as ?s or with the integer 0 (not 2^0).}
}
\value{
An integer specifying the minimum number of steps that the character must contain
}
\description{
Smallest number of steps that a character can take on any tree
}
