% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_prepare_data.R
\name{MorphyDat}
\alias{MorphyDat}
\title{PhyDat to MorphyDat}
\usage{
MorphyDat(phydat)
}
\arguments{
\item{phydat}{A \code{phyDat} object}
}
\value{
This function returns a matrix of class \code{*phyDat}.
   \itemize{
\item     Each row of the matrix corresponds to an operational taxonomic unit;
\item     Each column of the matrix represents a transformation series, with all transformation
          series containing three or more inapplicable tokens (and thus requiring the use of the 
          modified Fitch algorithm) listed first;
\item     Each cell in the matrix contains an integer that denotes the possible tokens.
          This integer is the decimal representation of the binary figure represented by a row
          in the contrast matrix (in reverse order), after
          phangorn:::prepareFitchData.          
  }
  In addition to the standard \link{phyDat} attributes, the returned object bears an attribute
  \code{inapp.level}, specifying the level that corresponds to the inapplicable token, \code{
  inapp.chars}, specifying the number of transformation series in which three or more inapplicable
  tokens are present (thus requiring the use of the modified Fitch algorithm), and [#TODO]
  \code{min.steps}, specifying the minimum number of steps necessary to obtain the observed
  distribution of tokens (necessary when calculating implied weights).
}
\description{
Converts a PhyDat object to allow processing by MorphyDat
}
\examples{
data(inapplicable.datasets)
morphy <- MorphyDat(inapplicable.phyData[[1]])

}
\seealso{
\code{\link{phyDat}}
}
\author{
Martin R. Smith
}
