% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{MorphyErrorCheck}
\alias{MorphyErrorCheck}
\title{Check for error whilst modifying Morphy object}
\usage{
MorphyErrorCheck(action)
}
\arguments{
\item{action}{action to perform}
}
\description{
Check for error whilst modifying Morphy object
}
\keyword{internal}
