% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_tree_rearrange.R
\name{MorphyRearrangeTree}
\alias{MorphyRearrangeTree}
\alias{MorphyRearrange}
\title{Rearrange phylogenetic tree}
\usage{
MorphyRearrangeTree(tree, morphyObj, Rearrange, minScore = NULL,
  returnSingle = TRUE, iter = "?", cluster = NULL, verbosity = 0L)

MorphyRearrange(parent, child, morphyObj, inputScore = 1e+07, hits = 0,
  EdgeSwapper, minScore = NULL, returnSingle = TRUE, iter = "?",
  cluster = NULL, verbosity = 0L)
}
\arguments{
\item{tree}{a rooted bifurcating phylogenetic tree with the desired outgroup, with its labels
in an order that matches the Morphy object, and the attributes
\code{score}, the tree's optimality score, and 
\code{hits}, the number of times the best score has been hit in the calling function.}

\item{morphyObj}{A morphy object, perhaps created with \code{\link{PhyDat2Morphy}}.}

\item{Rearrange}{a rearrangement function that returns a tree: probably one of 
\code{\link{RootedNNI}}, \code{\link{RootedSPR}} or
\code{\link{RootedTBR}}.}

\item{minScore}{trees longer than \code{min.score}, probably the score of the starting tree,
will be discarded.}

\item{returnSingle}{returns all trees if \kbd{FALSE} or a randomly selected tree if \kbd{TRUE}.}

\item{iter}{iteration number of calling function, for reporting to user only.}

\item{cluster}{a cluster, prepared with code{link{PrepareCluster}}, to accelerate 
searches on multicore machines. #TODO not tested. #TODO restore code link}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{parent}{the first column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 1]}

\item{child}{the second column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 2]}

\item{inputScore}{the score of the tree, if known}

\item{hits}{number of times that this score has been hit}

\item{RearrangeEdges}{a function that rearranges a parent and child vector, 
and returns a list with modified vectors; for example
\code{\link{SPRSwap}}.}
}
\value{
{This function returns the most parsimonious of the trees generated, with attributes \code{hits} and \code{score}
 as described for argument \code{tree}, and with tip labels ordered to match morphyObj.}

a rearranged edgeList.
}
\description{
Rearrange phylogenetic tree
}
\details{
\code{MorphyRearrangeTree} performs one tree rearrangement of a specified type
}
\section{Functions}{
\itemize{
\item \code{MorphyRearrange}: optimised version that requires parent and child vectors to be extracted from a tree
}}

\seealso{
\itemize{
    \item \code{\link{RootedNNI}}
    \item \code{\link{RootedSPR}}
    \item \code{\link{RootedTBR}}
  }
}
\author{
Martin R. Smith

Martin R. Smith
}
\keyword{internal}
