% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_tree_search.R
\name{MorphyTreeSearch}
\alias{MorphyTreeSearch}
\title{Tree Search}
\usage{
MorphyTreeSearch(edgeList, morphyObj, Rearrange, maxIter = 100,
  maxHits = 20, stopAtScore = NULL, forestSize = 1L, cluster = NULL,
  verbosity = 1L, ...)
}
\arguments{
\item{edgeList}{a list containing the following:
- vector of integers corresponding to the parent of each edge in turn
- vector of integers corresponding to the child of each edge in turn
- (optionally) score of the tree
- (optionally, if score provided) number of times this score has been hit}

\item{morphyObj}{A morphy object, perhaps created with \code{\link{PhyDat2Morphy}}.}

\item{Rearrange}{Function to use to rearrange trees; example: 
\code{TreeSearch::\link[TreeSearch]{RootedTBR}}.}

\item{maxIter}{maximum iterations to conduct.}

\item{maxHits}{stop search after this many hits.}

\item{stopAtScore}{stop search as soon as this score is hit or beaten.}

\item{forestSize}{how many trees to hold.}

\item{cluster}{a cluster, prepared with code{link{PrepareCluster}}, to accelerate 
searches on multicore machines. #TODO not tested. #TODO restore code link}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{\dots}{additional variables to pass to \code{\link{MorphyRearrangeTree}}.}
}
\description{
Performs a tree search using a Morphy object
}
\details{
Does the hard work of searching for a most parsimonious tree, given the
parent and child vectors of a tree arranged in preorder (perhaps with 
\code{\link[TreeSearch]{RenumberEdges}}).
End-users are expected to access this function through its wrapper, TreeSearch
It is also called directly by MorphyRatchet and Sectorial functions
}
\author{
Martin R. Smith
}
\keyword{internal}
