% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{MorphyWeights}
\alias{MorphyWeights}
\title{Report the character weightings associated with a Morphy object}
\usage{
MorphyWeights(morphyObj)
}
\arguments{
\item{morphyObj}{A morphy object, perhaps created with \code{\link{PhyDat2Morphy}}.}
}
\value{
a matrix of dimensions (2, number of characters); row 1 lists the
exact rates specified by the user; row 2 the approximate (integral)
weights used by MorphyLib
}
\description{
Report the character weightings associated with a Morphy object
}
\author{
Martin R. Smith
}
