% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_apply_tipdata}
\alias{mpl_apply_tipdata}
\title{Commits parameters prior to nodal set calculations.}
\usage{
mpl_apply_tipdata(morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}
}
\value{
A Morphy error code.
}
\description{
Once the caller is satisfied with the setup of types, weights,
and partitioning, this function must be called, thereby committing the
parameters until any changes are made. If no character types have been
assigned, the function will fail with an error code.
}
\author{
Martin Brazeau
}
\keyword{internal}
