% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_attach_rawdata}
\alias{mpl_attach_rawdata}
\title{Attach raw character state data (i.e. tip data).}
\usage{
mpl_attach_rawdata(rawdata, morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}

\item{rawmatrix}{C-style string corresponding to the tip data for each taxon in turn.}
}
\value{
Morphy error code.
}
\description{
Attaches a raw data character state matrix in the form of a C-style
(i.e. NULL-terminated) string. This can be the matrix block extracted from a
Nexus file or an xread table format. The matrix should contain no terminal or
tip labels.
}
\author{
Martin Brazeau
}
\keyword{internal}
