% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_delete_Morphy}
\alias{mpl_delete_Morphy}
\title{Destroys an instance of a Morphy object.}
\usage{
mpl_delete_Morphy(morphyobj)
}
\arguments{
\item{morphyobj}{A Morphy object to be destroyed.}
}
\value{
A Morphy error code.
}
\description{
Destroys an instance of the Morphy object, calling all
destructor for internal object completely returning the memory to the system.
}
\author{
Martin Brazeau
}
\keyword{internal}
