% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_delete_rawdata}
\alias{mpl_delete_rawdata}
\title{Deletes the caller-input data.}
\usage{
mpl_delete_rawdata(morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}
}
\value{
Morphy error code.
}
\description{
Deletes all of the user-input data and restores all parameters
to their original values, except for the dimensions of the matrix.
}
\author{
Thomas Guillerme
}
\keyword{internal}
