% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_first_down_recon}
\alias{mpl_first_down_recon}
\title{Reconstructs the first (downpass) nodal reconstructions}
\usage{
mpl_first_down_recon(node_id, left_id, right_id, morphyobj)
}
\arguments{
\item{node_id}{The index of the node being reconstructed.}

\item{left_id}{The index of the left descendant.}

\item{right_id}{The index of the right descendant.}

\item{morphyobj}{An instance of the Morphy object.}
}
\value{
The integral parsimony length (right now)
}
\description{
Reconstructs the preliminary nodal set for all characters for a 
particular node. This function is called over a postorder sequence of internal 
nodes where left and right descendants are known.
Because this function needs to be fairly high-performance, it does not do much 
checking for parameter validity, thus unsafe usage of this function might not
be caught. It is up to calling functions to ensure that the appropriate 
parameters have been set before use.
}
\author{
Martin Brazeau
}
\keyword{internal}
