% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_get_charac_weight}
\alias{mpl_get_charac_weight}
\title{Retrieve the weight of a character in the dataset}
\usage{
mpl_get_charac_weight(charID, morphyobj)
}
\arguments{
\item{charID}{Number of the character (i.e. first character is number 1)}

\item{morphyobj}{An instance of the Morphy object.}
}
\value{
A list, detailing (item 1) the exact weight of the character; (item 2) the integer
approximation used by Morphy.
}
\description{
Gets the weights of a character in the dataset.
}
\author{
Martin R. Smith
}
\keyword{internal}
