% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_get_num_charac}
\alias{mpl_get_num_charac}
\title{Retrieve the number of character (columns) in the dataset.}
\usage{
mpl_get_num_charac(morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}
}
\value{
The number of internal nodes.
}
\description{
Retrieves the number of character (columns) in the dataset.
}
\author{
Martin Brazeau
}
\keyword{internal}
