% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_get_num_internal_nodes}
\alias{mpl_get_num_internal_nodes}
\title{Gets the number of internal nodal reconstruction sets being used by
MorphyLib.}
\usage{
mpl_get_num_internal_nodes(morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}
}
\value{
The number of internal nodes.
}
\description{
Gets the number of internal nodal reconstruction sets being used
by MorphyLib.
}
\author{
Martin Brazeau
}
\keyword{internal}
