% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_get_numtaxa}
\alias{mpl_get_numtaxa}
\title{Retrieve the number of taxa (rows) in the dataset.}
\usage{
mpl_get_numtaxa(morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}
}
\value{
The number of taxa if success, otherwise an error code.
}
\description{
Retrieves the number of taxa (rows) in the dataset.
}
\author{
Martin Brazeau
}
\keyword{internal}
