% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_init_Morphy}
\alias{mpl_init_Morphy}
\title{Sets up the dimensions of the dataset.}
\usage{
mpl_init_Morphy(numtaxa, numchars, morphyobj)
}
\arguments{
\item{morphyobj}{An instance of the Morphy object.}

\item{ntax}{The number of taxa (or tips/terminals).}

\item{nchar}{The number of characters (i.e. transformation series) in the
data set.}
}
\value{
Morphy error code.
}
\description{
Provides initial dimensions for the dataset, which will
constrain any input matrix supplied to Morphy.
}
\author{
Martin Brazeau
}
\keyword{internal}
