% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_new_Morphy}
\alias{mpl_new_Morphy}
\title{Creates a new instance of a Morphy object}
\usage{
mpl_new_Morphy()
}
\value{
A void pointer to the Morphy instance. NULL if unsuccessful.
}
\description{
Creates a new empty Morphy object. All fields are unpopulated
and uninitialised.
}
\examples{
morphyObj <- mpl_new_Morphy() # Create new object
## Do some stuff ... ##
mpl_delete_Morphy(morphyObj) # Delete when done

}
\author{
Martin Brazeau
}
\keyword{internal}
