% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_set_charac_weight}
\alias{mpl_set_charac_weight}
\title{Set the weight of a character in the dataset}
\usage{
mpl_set_charac_weight(charID, weight, morphyobj)
}
\arguments{
\item{charID}{Number of the character (i.e. first character is number 1)}

\item{weight}{Weight to assign}

\item{morphyobj}{An instance of the Morphy object.}
}
\value{
An error code.
}
\description{
Sets the weight of a character in the dataset.
}
\author{
Martin R. Smith
}
\keyword{internal}
