% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_translate_error}
\alias{mpl_translate_error}
\title{Converts a numeric error code to human-readable format}
\usage{
mpl_translate_error(errorCode)
}
\arguments{
\item{errorCode}{Non-positive integer to be converted}
}
\value{
A character string corresponding to the provided error code
}
\description{
Converts a numeric error code to human-readable format
}
\examples{
mpl_translate_error(-1) # "ERR_INVALID_SYMBOL"

}
\author{
Martin R. Smith
}
\keyword{internal}
