% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphyex.R
\name{mpl_update_lower_root}
\alias{mpl_update_lower_root}
\title{Updates the nodal sets for a lower ('dummy') root node}
\usage{
mpl_update_lower_root(l_root_id, root_id, morphyobj)
}
\arguments{
\item{l_root_id}{The index of the lower root.}

\item{root_id}{The index of the upper root node.}

\item{morphyobj}{An instance of the Morphy object.}
}
\value{
A Morphy error code.
}
\description{
If trees are rooted, then Morphy uppass functions
require a lower or 'dummy' root in order to function properly. This
function should be called to set the nodal state sets to the dummy
root. The nodal set will be equal to the set of the root node, unless
there is an ambiguous union of applicable and gap tokens when gaps are 
treated as in applicable. In which case, the set union is resolved in 
favour of any applicable tokens in the set.
}
\author{
Thomas Guillerme
}
\keyword{internal}
