% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{NamedConstant}
\alias{NamedConstant}
\title{Named constant}
\usage{
NamedConstant(X, name)
}
\arguments{
\item{X}{a vector.}

\item{name}{name to apply to the vector.}
}
\value{
the vector, named as requested
}
\description{
Named constant
}
\author{
Martin R. Smith
}
