% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_display.R
\name{NewickTree}
\alias{NewickTree}
\title{Newick Tree}
\usage{
NewickTree(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
A character string describing \code{tree} in Newick format
}
\description{
Writes a tree in Newick format
}
