% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{NeworderPruningwise}
\alias{NeworderPruningwise}
\title{neworder_pruningwise
Wrapper for the ape function}
\usage{
NeworderPruningwise(nTaxa, nb.node, parent, child, nb.edge)
}
\description{
neworder_pruningwise
Wrapper for the ape function
}
\keyword{internal}
