% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{NJTree}
\alias{NJTree}
\title{Neighbour Joining Tree}
\usage{
NJTree(dataset)
}
\arguments{
\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree.}
}
\value{
an object of class \code{phylo}
}
\description{
Generates a rooted neighbour joining tree, with no edge lengths
}
\author{
Martin R. Smith
}
