% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNI.R
\name{NNI}
\alias{NNI}
\alias{RootedNNI}
\title{NNI}
\usage{
NNI(tree, edgeToBreak = NULL)

RootedNNI(tree, edgeToBreak = NULL)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{edgeToBreak}{(optional) integer specifying the index of an edge to bisect/prune, generated randomly if not specified}
}
\value{
Returns a tree with class \code{phylo}.
}
\description{
Nearest Neighbour Interchange

Rooted NNI
}
\details{
This function performs a single random \acronym{NNI} iteration.

Performs a single iteration of the nearest-neighbour interchange algorithm.
Based on the corresponding phangorn function, but re-coded to improve speed.

Branch lengths are not supported.
}
\section{Functions}{
\itemize{
\item \code{RootedNNI}: Perform \acronym{NNI} rearrangement, retaining position of root
}}

\examples{
tree <- ape:::rtree(20, br=NULL)
NNI(tree)

}
\references{
The algorithms are summarized in
Felsenstein, J. 2004. \cite{Inferring Phylogenies.} Sinauer Associates, Sunderland, Massachusetts.
}
\author{
Martin R. Smith
}
