% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{NonDuplicateRoot}
\alias{NonDuplicateRoot}
\title{Non-duplicate root}
\usage{
NonDuplicateRoot(parent, child, nEdge = length(parent))
}
\arguments{
\item{parent}{the first column of the edge matrix of a tree of class
\code{\link{phylo}}, i.e. \code{tree$edge[, 1]}}

\item{child}{the second column of the edge matrix of a tree of class
\code{\link{phylo}}, i.e. \code{tree$edge[, 2]}}

\item{nEdge}{(optional) integer specifying the number of edges of a tree of
class \code{\link{phylo}}, i.e. \code{dim(tree$edge)[1]}}
}
\description{
Identify, for each edge, whether it is not a duplicate of the root edge
}
\author{
Martin R. Smith
}
\keyword{internal}
