% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldSPR.R
\name{oldSPR}
\alias{oldSPR}
\title{Subtree Pruning and Rearrangement}
\usage{
oldSPR(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}
}
\description{
Perform one \acronym{SPR} rearrangement on a tree
}
\details{
Equivalent to phangorn's kSPR, but faster.
Note that rearrangements that only change the position of the root WILL be returned by 
\code{SPR}.  If the position of the root is irrelevant (as in Fitch parsimony, for example)
then this function will occasionally return a functionally equivalent topology.  
\code{RootIrrelevantSPR} will search tree space more efficiently in these cases.
}
\author{
Martin R. Smith
}
