% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneOverlap}
\alias{oneOverlap}
\title{One Overlap}
\format{An object of class \code{list} of length 100.}
\usage{
oneOverlap
}
\description{
Information content of a single overlap
Internal cache of all values for up to 100 terminals, called from within \code{MutualPhylogeneticInfoSplits}.
}
\details{
The benefits of cacheing decrease an nTerminals increases (as a match is less likely)
and the cost rises (in terms of storage space).
}
\keyword{datasets}
