% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{OrderEdgesNumberNodes}
\alias{OrderEdgesNumberNodes}
\title{Order edges and number nodes
Wrapper for the C function}
\usage{
OrderEdgesNumberNodes(parent, child, nTips, nEdge = length(parent))
}
\value{
an edge matrix for a tree following the usual convention for edge and node numbering
}
\description{
Order edges and number nodes
Wrapper for the C function
}
\keyword{internal}
