% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{PartitionInfo}
\alias{PartitionInfo}
\title{Information content of partitions within a tree}
\usage{
PartitionInfo(x)
}
\arguments{
\item{tree}{A tree of class \code{phylo}, a list of trees, or a \code{multiPhylo} object.}
}
\description{
Sums the phylogenetic information content for all partitions within a
phylogenetic tree.  This value will be greater than the total information
content of the tree where a tree contains multiple partitions, as
these partitions will contain mutual information
}
\author{
Martin R. Smith
}
\keyword{internal}
