% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{PhyDat2Morphy}
\alias{PhyDat2Morphy}
\title{Initialize a Morphy Object from a phyDat object}
\usage{
PhyDat2Morphy(phy)
}
\arguments{
\item{phy}{An object of class \code{\link{phyDat}}.}
}
\value{
A pointer to an initialized Morphy object.
}
\description{
Creates a new Morphy object with the same size and characters as the phyDat object
}
\author{
Martin R. Smith
}
