% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_prepare_data.R
\name{ConvertToString}
\alias{ConvertToString}
\alias{PhyToString}
\alias{ProfileToString}
\title{Extract character data from a phyDat object as a string}
\usage{
ConvertToString(phyByTaxon, phyLevels, phyChars, phyContrast, phyIndex, ps,
  byTaxon, concatenate)

PhyToString(phy, ps = "", useIndex = TRUE, byTaxon = TRUE,
  concatenate = TRUE)

ProfileToString(dataset, ps = "", useIndex = TRUE, byTaxon = TRUE,
  concatenate = TRUE)
}
\arguments{
\item{phyByTaxon}{a phyDat object or other list of character values, a taxon at a time}

\item{ps}{Character specifying text, perhaps ';', to append to the end of the string}

\item{byTaxon}{If TRUE, write one taxon followed by the next.
If FALSE, write one character followed by the next.}

\item{concatenate}{Logical specifying whether to concatenate all characters/taxa
into a single string, or to return a separate string
for each entry.}

\item{phy}{An object of class \code{\link{phyDat}}}

\item{useIndex}{(default: TRUE) Print duplicate characters multiple 
times, as they appeared in the original matrix}
}
\description{
Extract character data from a phyDat object as a string
}
\section{Functions}{
\itemize{
\item \code{ConvertToString}: Generic underlying function

\item \code{ProfileToString}: Convert from profileDat object.
}}

\author{
Martin R. Smith
}
\keyword{internal}
