% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{PrepareCluster}
\alias{PrepareCluster}
\title{Prepare a cluster for use in tree search functions}
\usage{
PrepareCluster(cores)
}
\arguments{
\item{cores}{Number of cores to include in the cluster.}
}
\value{
Returns a reference to a cluster.
}
\description{
\code{PrepareCluster} creates a cluster of multiple cores and prepares it to analyse phylogenetic trees.
}
\examples{
\dontrun{
  PrepareCluster(4)
}

}
\author{
Martin R. Smith
}
