% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTable.R
\name{print.ClusterTable}
\alias{print.ClusterTable}
\title{Print \code{ClusterTable} object}
\usage{
\method{print}{ClusterTable}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ClusterTable}.}

\item{\dots}{Additional arguments for consistency with S3 method (unused).}
}
\description{
S3 method for objects of class \code{ClusterTable}.
}
