% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomTreeScore.R
\name{RandomMorphyTree}
\alias{RandomMorphyTree}
\title{Random postorder tree}
\usage{
RandomMorphyTree(nTip)
}
\arguments{
\item{nTip}{number of tips (minimum 3)}
}
\value{
A list with three elements, each a vector of integers, respectively containing:
- The parent of each tip and node, in order
- The left child of each node
- The right child of each node.
}
\description{
Random postorder tree
}
