% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{RandomPostorder}
\alias{RandomPostorder}
\title{Random postorder tree}
\usage{
RandomPostorder(nTip)
}
\arguments{
\item{nTip}{number of tips}
}
\value{
a list of three integer vectors: 
         First entry: parentOf: For each node, numbered in postorder, the number of its parent node.
         Second entry: leftChild: For each internal node, numbered in postorder, the number of its left 
                  child node or tip.
         Third entry: rightChild: For each internal node, numbered in postorder, the number of its right
                  child node or tip.
}
\description{
Random postorder tree
}
