% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratchet.R
\name{Ratchet}
\alias{Ratchet}
\title{Parsimony Ratchet}
\usage{
Ratchet(tree, data, TreeScorer = FitchScore, returnAll = FALSE,
  rooted = TRUE, ratchIter = 100, searchIter = 2000, searchHits = 40,
  ratchHits = 10, verbosity = 0, rearrangements = "NNI",
  suboptimal = 1e-08, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{data}{a dataset in the format required by TreeScorer}

\item{returnAll}{Set to \code{TRUE} to report all MPTs encountered during the search, perhaps to analyze consensus}

\item{rooted}{whether to retain the position of the root in tree search (TRUE by default)}

\item{verbosity}{verbosity of output written to terminal; higher values print more messages}

\item{rearrangements}{method(s) to use when rearranging trees. Provide:
the character string "TBR" to conduct TBR, SPR and NNI rearrangements, or "TBR only"
to just perform TBR rearrangements (retaining the position of the root if \code{outgroup = TRUE})
OR: a list of functions to use, one at a time, as the \code{Rearrange} parameter
    in successive calls to DoTreeSearch}

\item{\dots}{further arguments to pass to \code{TreeScorer} function (e.g. \code{\link{TipsAreColumns}})}

\item{concavity}{concavity constant for implied weighting (not currently implemented!)}

\item{maxit}{maximum ratchet iterations to perform;}

\item{maxIter}{maximum rearrangements to perform on each bootstrap or ratchet iteration;}

\item{maxHits}{maximum times to hit best score before terminating a tree search within a pratchet iteration;}

\item{k}{stop when k ratchet iterations have found the same best score;}
}
\value{
This function returns a tree modified by parsimony ratchet iteration, retaining the position of the root.
}
\description{
\code{Ratchet} uses the parsimony ratchet (Nixon 1999) to search for a more parsimonious tree.
}
\examples{
Ratchet(RandomTree(Lobo.phy), Lobo.phy, rooted=TRUE)

}
\references{
Nixon, K. C. (1999). \cite{The Parsimony Ratchet, a new method for rapid parsimony analysis.} Cladistics, 15(4), 407-414. doi:\href{http://dx.doi.org/10.1111/j.1096-0031.1999.tb00277.x}{10.1111/j.1096-0031.1999.tb00277.x}
}
\seealso{
\code{\link{Ratchet}}

\code{\link{TreeSearch}}

\code{\link{SectorialSearch}}
}
\author{
Martin R. Smith

Adapted from \code{\link[phangorn]{pratchet}} in the \pkg{phangorn} package, which does not preserve the position of the root.
}
\keyword{tree}
