% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_tree_search.R
\name{RatchetSearch}
\alias{RatchetSearch}
\alias{RatchetConsensus}
\title{Parsimony Ratchet}
\usage{
RatchetSearch(tree, dataset, keepAll = FALSE, maxIt = 100, maxIter = 5000,
  maxHits = 40, k = 10, stopAtScore = NULL, verbosity = 1L,
  rearrangements = list(TBRCore, SPRCore, NNICore), ...)

RatchetConsensus(tree, dataset, maxIt = 5000, maxIter = 500, maxHits = 20,
  k = 10, verbosity = 0L, rearrangements = list(RootedNNI),
  nSearch = 10, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree.}

\item{keepAll}{Set to \kbd{TRUE} to report all MPTs encountered during the search, perhaps to analyze consensus}

\item{maxIt}{maximum ratchet iterations to perform;}

\item{maxIter}{maximum rearrangements to perform on each bootstrap or ratchet iteration;}

\item{maxHits}{maximum times to hit best score before terminating a tree search within a pratchet iteration;}

\item{k}{stop when k ratchet iterations have found the same best score;}

\item{stopAtScore}{stop search as soon as this score is hit or beaten.}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{rearrangements}{(list of) function(s) to use when rearranging trees
e.g. \code{list(TreeSearch::RootedTBR, TreeSearch::RootedNNI)}}

\item{\dots}{other arguments to pass to subsequent functions.}

\item{nSearch}{Number of Ratchet searches to conduct (for RatchetConsensus)}
}
\value{
This function returns a tree modified by parsimony ratchet iteration, retaining the position of the root.
}
\description{
\code{RatchetSearch} uses the parsimony ratchet (Nixon 1999) to search for a more parsimonious tree.
}
\section{Functions}{
\itemize{
\item \code{RatchetConsensus}: returns a list of optimal trees produced by nSearch Ratchet searches
}}

\examples{
{
data('inapplicable.datasets')
my.phyDat <- inapplicable.phyData[[1]]
RatchetSearch(tree=TreeSearch::RandomTree(my.phyDat, root=names(my.phyDat)[1]), 
        dataset=my.phyDat, maxIt=1, maxIter=50)
}
}
\references{
Nixon, K. C. (1999). \cite{The Parsimony Ratchet, a new method for rapid parsimony analysis.}
 Cladistics, 15(4), 407-414. doi:\href{http://dx.doi.org/10.1111/j.1096-0031.1999.tb00277.x}{10.1111/j.1096-0031.1999.tb00277.x}
}
\seealso{
\code{\link[phangorn]{pratchet}}

\code{\link{BasicSearch}}
}
\author{
Martin R. Smith

Adapted from \code{\link[phangorn]{pratchet}} in the \pkg{phangorn} package, which does not preserve the position of the root.
}
\keyword{tree}
