% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{RearrangeTree}
\alias{RearrangeTree}
\alias{RearrangeEdges}
\title{Rearrange phylogenetic tree}
\usage{
RearrangeTree(tree, TreeScorer = FitchScore, Rearrange = RootedNNI,
  minScore = NULL, returnSingle = TRUE, iter = "<unknown>",
  verbosity = 0, ...)

RearrangeEdges(parent, child, dataset, TreeScorer, inputScore = 1e+07,
  hits = 0, EdgeSwapper, minScore = NULL, returnSingle = TRUE,
  iter = "?", cluster = NULL, verbosity = 0L)
}
\arguments{
\item{tree}{a rooted bifurcating phylogenetic tree with the desired outgroup, with its labels
in an order that matches the Morphy object, and the attributes
\code{pscore}, the tree's parsimony score, and 
\code{hits}, the number of times the best score has been hit in the calling function;}

\item{TreeScorer}{function by which trees should be scored; default \code{\link{FitchScore}}}

\item{Rearrange}{a rearrangement function: probably one of 
\code{\link{RootedNNI}}, \code{\link{RootedSPR}} or \code{\link{RootedTBR}};}

\item{minScore}{trees longer than \code{minScore}, probably the score of the starting tree,
will be discarded;}

\item{returnSingle}{returns all trees if \kbd{FALSE} or a randomly selected tree if \kbd{TRUE};}
\item{iter}{iteration number of calling function, for reporting to user only;}

\item{iter}{iteration number of calling function, for reporting to user only.}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{\dots}{further arguments to pass to \code{TreeScorer} function (e.g. \code{\link{TipsAreColumns}}, \code{dataset})}

\item{dataset}{A dataset; will be the third argument passed to TreeScorer}

\item{cluster}{a cluster, prepared with code{link{PrepareCluster}}, to accelerate 
searches on multicore machines. #TODO not tested. #TODO restore code link}

\item{tree}{a rooted bifurcating phylogenetic tree with the desired outgroup, with its labels
in an order that matches the Morphy object, and the attributes
\code{score}, the tree's optimality score, and 
\code{hits}, the number of times the best score has been hit in the calling function.}

\item{TreeScorer}{Function to score trees. Will be sent three parameters: parent, child and dataset.}

\item{RearrangeEdges}{a function that rearranges a parent and child vector, 
and returns a list with modified vectors; for example
\code{\link{SPRSwap}}.}

\item{minScore}{trees longer than \code{min.score}, probably the score of the starting tree,
will be discarded.}

\item{returnSingle}{returns all trees if \kbd{FALSE} or a randomly selected tree if \kbd{TRUE}.}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}
}
\value{
{This function returns the most parsimonious of the trees generated, with attributes \code{hits} and \code{pscore}
 as described for argument \code{tree}, and with tip labels ordered to match morphyObj.}
}
\description{
Rearrange phylogenetic tree
}
\details{
\code{RearrangeTree} performs one tree rearrangement of a specified type
}
\section{Functions}{
\itemize{
\item \code{RearrangeEdges}: Rearrange edges
}}

\examples{
data('Lobo')
random.tree <- RandomTree(Lobo.phy)
RearrangeTree(random.tree, dataset=Lobo.phy, Rearrange=RootedNNI)

}
\seealso{
\itemize{
    \item \code{\link{RootedNNI}}
    \item \code{\link{RootedSPR}}
    \item \code{\link{RootedTBR}}
  }
}
\author{
Martin R. Smith
}
