% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{referenceTree}
\alias{referenceTree}
\title{Tree topology for matrix simulation}
\format{A single phylogenetic tree saved as an object of class \code{phylo}}
\source{
\url{https://dx.doi.org/10.1111/pala.12236}
}
\usage{
referenceTree
}
\description{
The tree topology used to generate the matrices in \code{\link{congreveLamsdellMatrices}}
Congreve & Lamsdell (2016)
}
\examples{
  data(referenceTree)
  plot(referenceTree)

}
\references{
\insertRef{Congreve2016}{TreeSearch}
\insertRef{Congreve2016dd}{TreeSearch}
}
\keyword{datasets}
