% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{Renumber}
\alias{Renumber}
\title{Renumber a tree's nodes and tips}
\usage{
Renumber(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
This function returns a tree of class \code{phylo}
}
\description{
\code{Renumber} numbers the nodes and tips in a tree to conform with the
phylo standards.
}
\examples{
library('ape')
tree <- rtree(10)
Renumber (tree)

}
\author{
Martin R. Smith
}
