% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{ReportMatching}
\alias{ReportMatching}
\title{List matched clades as text}
\usage{
ReportMatching(clades1, clades2, taxonNames)
}
\arguments{
\item{clades1, clades2}{Logical matrices with columns specifying membership
of each corresponding matched clade}

\item{taxonNames}{Character vector listing names of taxa corresponding to
each row in \code{clades#}}
}
\description{
List matched clades as text
}
\seealso{
VisualizeMatching
}
\keyword{internal}
