% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{RightmostCharacter}
\alias{RightmostCharacter}
\title{Rightmost character of string}
\usage{
RightmostCharacter(string, len = nchar(string))
}
\description{
Rightmost character of string
}
\author{
Martin R. Smith
}
\keyword{internal}
