% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_rf.R
\name{RobinsonFoulds}
\alias{RobinsonFoulds}
\alias{RobinsonFouldsSplits}
\title{Robinson-Foulds Distance}
\usage{
RobinsonFoulds(tree1, tree2, normalize = FALSE, reportMatching = FALSE)

RobinsonFouldsSplits(splits1, splits2, reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}
}
\description{
An inefficient implementation of the Robinson-Foulds distance, included
for use with \code{\link{VisualizeMatching}}.  To generate the RF distance efficiently,
use the function \code{\link{ape}{treedist}}.
}
\details{
Note that if \code{reportMatching = TRUE}, the \code{pairScores} attribute returns
a logical matrix specifying whether each pair of partitions is identical.
}
\section{Functions}{
\itemize{
\item \code{RobinsonFouldsSplits}: Calculate Robinson-Foulds distance from splits
instead of trees.
}}

\section{Normalization}{


Normalized against the total number of partitions that are present.
}

\references{
\insertRef{Robinson1981}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{KendallColijn}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith
}
\concept{tree distances}
