% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_rf.R
\name{RobinsonFouldsInfo}
\alias{RobinsonFouldsInfo}
\alias{RobinsonFoulds}
\title{(Information-adjusted) Robinson-Foulds distance}
\usage{
RobinsonFouldsInfo(tree1, tree2, similarity = FALSE, normalize = FALSE,
  reportMatching = FALSE)

RobinsonFoulds(tree1, tree2, similarity = FALSE, normalize = FALSE,
  reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{similarity}{Logical specifying whether to report the result as a tree
similarity, rather than a difference}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}
}
\description{
Calculates the Robinson-Foulds distance, or the equivalent similarity
measure, optionally annotating matched partitions and weighting partitions
according to their phylogenetic information content.
}
\details{
\code{RobinsonFoulds} is an inefficient implementation of the Robinson-Foulds
distance, included for use with \code{\link{VisualizeMatching}}.
To generate the RF distance efficiently,
use the function \code{\link{ape}{treedist}}.

Note that if \code{reportMatching = TRUE}, the \code{pairScores} attribute returns
a logical matrix specifying whether each pair of partitions is identical.

\code{RobinsonFouldsInfo} calculates the tree similarity or distance by summing
the phylogenetic information content of all partitions that are (or are not)
identical in both trees.  Consequently, partitions that are more likely
to be identical by chance alone make a smaller contribution to overall
tree distance, because their similarity is less remarkable.
}
\section{Functions}{
\itemize{
\item \code{RobinsonFoulds}: An inefficient implementation of the Robinson-Foulds distance, included
for use with \code{\link{VisualizeMatching}}.  To generate the RF distance efficiently,
use the function \code{\link{ape}{treedist}}.
}}

\section{Normalization}{

\itemize{
\item \code{RobinsonFoulds} is normalized against the total number of partitions that
are present.
\item \code{RobinsonFouldsInfo} is normalized against the sum of the phylogenetic
information of all splits in both trees, treated independently.
}
}

\references{
\insertRef{Robinson1981}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{KendallColijn}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith
}
\concept{tree distances}
