% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNI.R
\name{RootedNNI}
\alias{RootedNNI}
\title{Rearrange a rooted tree}
\usage{
RootedNNI(tree)
RootedSPR(tree)
RootedTBR(tree)
}
\arguments{
\item{tree}{A bifurcating tree of class \code{\link{phylo}}, with all nodes resolved}

\item{edgeToBreak}{(optional) integer specifying which edge should be broken (pruned)}

\item{mergeEdges}{(optional) vector of length 1 or 2, listing edge(s) to be joined:
In SPR, this is where the pruned subtree will be reconnected.
In TBR, these edges will be reconnected (so must be on opposite
sides of \code{edgeToBreak}); if only a single edge is specified,
the second will be chosen at random}
}
\value{
This function returns a tree, in \code{phylo} format.
}
\description{
This function performs a rearrangement iteration on a tree, retaining the position of the root.
}
\details{
A single \acronym{NNI}, \acronym{SPR} or \acronym{TBR} rearrangement is performed, subject to the constraint that 
no taxon may be moved to the opposite side of the root node.
Branch lengths are not (yet) supported.
}
\examples{
{
  require('ape')
  tree <- read.tree(text='(((a,b),c),(d,(e,f)));')
  tree <- SetOutgroup(tree, c('e', 'f'))
  plot(tree)
  dev.new()
  plot(RootedNNI(tree))
  plot(RootedSPR(tree))
  plot(RootedTBR(tree))
}


}
\seealso{
\itemize{
\item \code{\link{SetOutgroup}}, set the outgroup of the phylogenetic tree
\item \code{\link{NNI}}, unrooted \acronym{NNI} and \acronym{SPR}
\item \code{\link{TBR}}, unrooted \acronym{TBR}
}
}
\author{
Martin R. Smith
\code{RootedNNI} is abridged from the \pkg{phangorn} function \code{nnin}
}
