% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{RootTree}
\alias{RootTree}
\title{Root Tree on specified tips}
\usage{
RootTree(tree, outgroupTips)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}

\item{outgroupTips}{Character vector specifying the names of the tips to
include in the outgroup.}
}
\value{
A tree of class phylo, rooted on the smallest clade that contains the specified tips
}
\description{
Roots a tree on the smallest clade containing the specified tips.
}
\author{
Martin R. Smith
}
