% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_tree.R
\name{Sample1}
\alias{Sample1}
\title{Quick sample}
\usage{
Sample1(x, len = length(x))
}
\arguments{
\item{x}{vector to sample}

\item{len}{length of vector}
}
\description{
Faster than inbuilt sample because it avoids some checks
}
\keyword{internal}
