% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{SampleOne}
\alias{SampleOne}
\title{Quick sample}
\usage{
SampleOne(x, len = length(x))
}
\arguments{
\item{x}{vector to sample}

\item{len}{length of vector}
}
\description{
Faster than inbuilt sample because it avoids some checks
}
\keyword{internal}
