% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sectorial.R
\name{SectorHasData}
\alias{SectorHasData}
\title{Sector Data
Check that chosen sector contains parsimony-informative data}
\usage{
SectorHasData(dataset, tips)
}
\arguments{
\item{dataset}{the dataset to subsample}

\item{tips}{character vector listing tips that exist in the sector}
}
\description{
The function simply checks that some characters have more than one state.
It's crude, but the cost of a false positive is low.
}
\keyword{internal}
