% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_mutual_information.R
\name{SelfHMI}
\alias{SelfHMI}
\title{Self-Hierarchical Mutual Information}
\usage{
SelfHMI(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo}, or a hierarchical partition list.}
}
\value{
A numeric value representing the self-hierarchical mutual information.
}
\description{
Calculate the self-hierarchical mutual information of a single tree.
This represents the total hierarchical information content of the tree.
}
\examples{
\dontrun{
library("TreeTools", quietly = TRUE)

bal6 <- BalancedTree(6)
SelfHMI(bal6)
}

}
\seealso{
Other tree distances: 
\code{\link{AHMI}()},
\code{\link{EHMI}()},
\code{\link{HierarchicalMutualInfo}()},
\code{\link{JaccardRobinsonFoulds}()},
\code{\link{KendallColijn}()},
\code{\link{MASTSize}()},
\code{\link{MatchingSplitDistance}()},
\code{\link{NNIDist}()},
\code{\link{NyeSimilarity}()},
\code{\link{PathDist}()},
\code{\link{Robinson-Foulds}},
\code{\link{SPRDist}()},
\code{\link{TreeDistance}()}
}
\concept{tree distances}
