% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{SetMorphyWeights}
\alias{SetMorphyWeights}
\title{Set the character weightings associated with a Morphy object}
\usage{
SetMorphyWeights(weight, morphyObj, checkInput = TRUE)
}
\arguments{
\item{weight}{A vector listing the new weights to be applied to each character}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}.}

\item{checkInput}{Whether to sanity-check input data before applying. 
Defaults to TRUE to protect the user from crashes.}
}
\value{
The Morphy error code generated when applying tipData
}
\description{
Set the character weightings associated with a Morphy object
}
\author{
Martin R. Smith
}
