% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{SetOutgroup}
\alias{SetOutgroup}
\title{SetOutgroup}
\usage{
SetOutgroup(tree, outgroup)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with all nodes resolved;}

\item{outgroup}{a vector of mode numeric or character specifying the new outgroup.}
}
\value{
This function returns a rooted tree with the new outgroup.
}
\description{
Root a phylogenetic tree
}
\details{
Sets the root of a phylogenetic tree, such that one child of the root node is \code{outgroup}.
}
\examples{
{
  require('ape')
  tree <- read.tree(text='(((a,b),c),(d,e));')
  plot(tree)
  plot(Root(tree, c('a', 'b')))
  plot(Root(tree, 3))
}

}
\seealso{
root ought to produce the same result as this function,
 but does not always do so in practice.
}
\author{
Martin R. Smith
}
