% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sevenTipDistances}
\alias{sevenTipDistances}
\title{Distances between unrooted seven-tip trees}
\format{An object of class \code{list} of length 10.}
\usage{
sevenTipDistances
}
\description{
Distances between each possible pairing of the 945 unrooted seven-tip trees
(equivalent to rooted 6-tip trees).  Following Kendall and Colijn (2016).
}
\references{
\insertRef{@Kendall2016}{TreeDist}
}
\keyword{datasets}
