% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{SingleCharMorphy}
\alias{SingleCharMorphy}
\title{Morphy object from single character}
\usage{
SingleCharMorphy(char)
}
\arguments{
\item{char}{state of each character at each tip in turn, in a format that will be converted
to a character string by \code{\link{paste0}(char, ';', collapse='')}.}
}
\value{
a pointer to a morphyObj. Don't forget to unload it when you're done with it:
        \code{morphyObj <- \link{UnloadMorphy}(morphyObj)}.
}
\description{
Morphy object from single character
}
\author{
Martin R. Smith
}
