% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_display.R
\name{SortTree}
\alias{SortTree}
\title{Sort tree}
\usage{
SortTree(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
A tree of class phylo, with each node sorted such that the larger clade is first.
}
\description{
Sorts each node into a consistent order, so similar trees look visually similar.
}
\author{
Martin R. Smith
}
