% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{SplitsCompatible}
\alias{SplitsCompatible}
\title{Are splits compatible?}
\usage{
SplitsCompatible(split1, split2)
}
\arguments{
\item{split1, split2}{Logical vectors listing terminals in same order, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}
}
\value{
\code{SplitsCompatible} returns a logical specifying whether the splits
provided are compatible with one another.
}
\description{
Splits are compatible if they are concave; i.e. they can both be true
simultaneously.
}
\author{
Martin R. Smith
}
