% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{SplitsRepeated}
\alias{SplitsRepeated}
\title{Are Splits Repeated?}
\usage{
SplitsRepeated(original, matches)
}
\arguments{
\item{original, matches}{Logical matrix describing the bipartition splits
defined by a tree (see \link{Tree2Splits} for format).  The first row
of \code{original} must be \code{FALSE} for all splits.}
}
\value{
\code{SplitsRepeated} returns a logical vector of length \code{ncol(original)},
specifying whether each split listed in \code{original} also appears in \code{matches}.
}
\description{
Determines whether any of the splits described in \code{original} are present in
\code{matches}.
}
\author{
Martin R. Smith
}
\keyword{internal}
