% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{StopUnlessBifurcating}
\alias{StopUnlessBifurcating}
\title{Check that all nodes in a tree are bifurcating.}
\usage{
StopUnlessBifurcating(parent)
}
\arguments{
\item{parent}{the first column of the edge matrix of a tree of class
\code{\link{phylo}}, i.e. \code{tree$edge[, 1]}}
}
\value{
Returns \code{NULL}, but will \code{stop} with an error message if a tree
does not appear to be bifurcating.
}
\description{
Check that all nodes in a tree are bifurcating.
}
\author{
Martin R. Smith
}
\keyword{internal}
