% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessiveApproximations.R
\name{SuccessiveWeights}
\alias{SuccessiveWeights}
\title{Successive Weights}
\usage{
SuccessiveWeights(tree, dataset)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree.}
}
\value{
Score of a tree, given the weighting instructions specified in the attributes of the dataset
}
\description{
Calculate weight for tree scored by successive approximations
}
\keyword{internal}
