% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{SupportColour}
\alias{SupportColour}
\alias{SupportColor}
\title{Support colour}
\usage{
SupportColour(support, show1 = TRUE)

SupportColor(support, show1 = TRUE)
}
\arguments{
\item{support}{A vector of doubles in the range 0-1}

\item{show1}{Logical specifying whether to display values of 1.
A transparent white will be returned if \code{FALSE}.}
}
\value{
A string containing the hexadecimal code for a colour picked from a
diverging scale, or \code{red} if a value is invalid.
}
\description{
Support colour
}
\section{Functions}{
\itemize{
\item \code{SupportColor}: alternative spelling
}}

