% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_tree.R
\name{RootedTBR}
\alias{RootedTBR}
\alias{TBR}
\title{Rooted TBR}
\usage{
RootedTBR(tree)

TBR(tree, edge.to.break = NULL)
}
\arguments{
\item{tree}{A bifurcating tree of class \code{\link{phylo}}, with all nodes resolved;}

\item{edge.to.break}{the index of an edge to bisect, generated randomly if not specified.}
}
\value{
This function returns a tree in \code{phyDat} format that has undergone one \acronym{TBR} iteration.
}
\description{
Rooted TBR

Tree bisection and reconnection
}
\details{
\code{TBR} performs a single random \acronym{TBR} iteration.

Branch lengths are not (yet) supported.
}
\section{Functions}{
\itemize{
\item \code{RootedTBR}: Perform \acronym{TBR} rearrangement, retaining position of root
}}

\examples{
{
library('ape')
tree <- rtree(20, br=NULL)
TBR(tree)
}
}
\references{
The \acronym{TBR} algorithm is summarized in
Felsenstein, J. 2004. \cite{Inferring Phylogenies.} Sinauer Associates, Sunderland, Massachusetts.
}
\seealso{
RootedTBR useful when the position of the root node should be retained.
}
\author{
Martin R. Smith
}
