% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partitions.R
\name{Tree2Splits}
\alias{Tree2Splits}
\alias{Tree2Bipartitions}
\title{Tree2Splits}
\usage{
Tree2Splits(tr)

Tree2Bipartitions(tr)
}
\arguments{
\item{tr}{A tree of class \code{\link[ape:read.tree]{phylo}}, with tips 
bearing integer labels (i.e. `tr$tip.label == 1:N`).}
}
\value{
Returns a two-dimensional array.  Columns correspond to unique
 bipartitions, named with the number of a node that denotes the partition.
 Rows correspond to tips `1:N`.
}
\description{
Converts a phylogenetic tree to an array of bipartition splits.
}
\examples{
Tree2Splits(ape::rtree(6, tip.label=1:6, br=NULL))

}
\author{
Martin R. Smith
}
\keyword{internal}
