% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{TreeDistanceReturn}
\alias{TreeDistanceReturn}
\title{Tree Distance Return}
\usage{
TreeDistanceReturn(pairScores, maximize = FALSE, reportMatching, splits1,
  splits2, taxonNames = NULL)
}
\arguments{
\item{pairScores}{Two-dimensional array listing the score of each possible
pairing of \code{splits1} (rows) with \code{splits2} (columns).}

\item{reportMatching}{Logical specifying whether to report details
of an optimal matching.}

\item{splits1, splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{taxonNames}{Character vector listing the names corresponding to each
row of \code{splits1} and \code{splits2}.
Corresponding parameters from within each function}
}
\description{
Generates the return value for Generalized Robinson-Foulds distances,
with (optionally) a report of the matching.
}
\author{
Martin R. Smith
}
\keyword{internal}
