% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{TreesMatchingSplit}
\alias{TreesMatchingSplit}
\alias{LogTreesMatchingSplit}
\title{Number of trees matching a bipartition split}
\usage{
TreesMatchingSplit(A, B)

LogTreesMatchingSplit(A, B)
}
\arguments{
\item{A, B}{Number of taxa in each partition.}
}
\description{
Calculates the number of unrooted bifurcated trees that are consistent with
a bipartition split that divides taxa into groups of size \code{A} and \code{B}.
}
\section{Functions}{
\itemize{
\item \code{LogTreesMatchingSplit}: Logarithm of the number of trees matching a split.
}}

\author{
Martin R. Smith
}
\concept{Split information}
