% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{TreeSpace}
\alias{TreeSpace}
\title{Interactive tree space analysis}
\usage{
TreeSpace()
}
\description{
Explore tree landscapes in shiny app.
}
\seealso{
Other tree space functions: 
\code{\link{ProjectionQuality}()},
\code{\link{SpectralClustering}()},
\code{\link{median.multiPhylo}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree space functions}
