% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{TwoTipTree}
\alias{TwoTipTree}
\title{TwoTipTree}
\usage{
TwoTipTree(tip1, tip2)
}
\arguments{
\item{tip1}{A character string representing the label for tip 1}

\item{tip2}{A character string representing the label for tip 2}
}
\value{
This function returns a \code{phylo} object with a single root node and two tips with
the specified labels.
}
\description{
Two-tipped tree
}
\details{
This function generates a tree of class \code{\link{phylo}} containing two tips.
}
\examples{
TwoTipTree('Homo', 'Pan')
}
\seealso{
SingleTaxonTree

bind.tree
}
\author{
Martin R. Smith
}
\keyword{tree}
