% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratchet.R
\name{UniqueExceptHits}
\alias{UniqueExceptHits}
\title{Unique trees (ignoring 'hits' attribute)}
\usage{
UniqueExceptHits(trees)
}
\description{
Unique trees (ignoring 'hits' attribute)
}
\author{
Martin R. Smith
}
\keyword{internal}
