% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_morphy_objects.R
\name{UnloadMorphy}
\alias{UnloadMorphy}
\title{Destroy a Morphy Object}
\usage{
UnloadMorphy(morphyObj)
}
\arguments{
\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}.}
}
\value{
Morphy error code, decipherable using \code{\link{mpl_translate_error}}
}
\description{
Best practice is to call \code{morphyObj <- UnloadMorphy(morphyObj)}
Failure to do so will cause a crash if UnloadMorphy is called on an object that 
has already been destroyed
}
\author{
Martin R. Smith
}
