% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{WithOneExtraStep}
\alias{WithOneExtraStep}
\title{Number of trees with one extra step}
\usage{
WithOneExtraStep(splits)
}
\arguments{
\item{splits}{A vector of integers listing the number of tips in each of a number of
tree splits (e.g. bipartitions).  For example, \kbd{c(3, 5)} states that
a character divides a set of eight tips into a group of three and a group
of five.}
}
\description{
Number of trees with one extra step
}
