% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{AllSplitPairings}
\alias{AllSplitPairings}
\title{All split pairings}
\usage{
AllSplitPairings(n)
}
\arguments{
\item{n}{Integer specifying the number of terminal taxa.}
}
\value{
\code{AllSplitPairings} returns a named vector, specifying the number of
split pairings producing
the variation of information given (in bits) in the name.  Splits
AB:CD and CD:AB are treated as distinct, so division of all values by four
is justified in cases where unique pairings only are required.
}
\description{
Calculates the variation of (clustering) information (Meila 2007)
for each possible pairing of non-trivial splits,
and tabulates the number of pairings with each similarity.
}
\details{
Trivial splits -- those that divide one or zero taxa from the rest --
are not considered.
}
\examples{
AllSplitPairings(5)

}
\references{
\itemize{
\item \insertRef{Meila2007}{TreeDist}
\item \insertRef{SmithDist}{TreeDist}
}
}
\author{
Martin R. Smith
}
