% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootstrapTree}
\alias{BootstrapTree}
\title{Bootstrap tree search with inapplicable data}
\usage{
BootstrapTree(tree, x, maxIter, maxHits, TreeScorer = FitchScore, track = 1,
  ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}

\item{x}{a dataset in the format required by TreeScorer}

\item{maxIter}{maximum number of iterations to perform in tree search}

\item{maxHits}{number of times to find optimal tree length before stopping tree search}

\item{track}{verbosity of output written to terminal; higher values print more messages}

\item{function}{by which trees should be scored; default \code{\link{FitchScore}}}
}
\value{
A tree that is optimal under a random sampling of the original characters
}
\description{
Bootstrap tree search with inapplicable data
}
