% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{Cladewise}
\alias{Cladewise}
\alias{Postorder}
\alias{Pruningwise}
\alias{Edgewise}
\title{Reorder tree Cladewise}
\usage{
Cladewise(tree, nTaxa = NULL, edge = tree$edge)

Postorder(tree, nTaxa = length(tree$tip.label), edge = tree$edge)

Pruningwise(tree, nTaxa = length(tree$tip.label), edge = tree$edge)

Edgewise(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{nTaxa}{(optional) number of tips in the tree}

\item{edge}{(optional) the value of tree$edge}
}
\value{
A tree with nodes numbered in postorder
}
\description{
A wrapper for \code{ape:::.reorder_ape}
}
\section{Functions}{
\itemize{
\item \code{Postorder}: Reorder tree in Postorder

\item \code{Pruningwise}: Reorder tree Pruningwise

\item \code{Edgewise}: Reorder tree 'edgewise' (special case of cladewise)
}}

\author{
Modified by Martin R. Smith from \code{.reorder_ape} in \pkg{ape} (Emmanuel Paradis)
}
\keyword{internal}
