% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPart.R
\name{clone}
\alias{clone}
\alias{clone.HPart}
\title{Clone / duplicate an object
\code{clone()} physically duplicates objects}
\usage{
clone(x, ...)

\method{clone}{HPart}(x, tipLabels = attr(x, "tip.label"), ...)
}
\arguments{
\item{x}{the object to be cloned}

\item{\dots}{additional parameters for methods}

\item{tipLabels}{Character vector specifying sequence in which to order
tip labels.}
}
\value{
\code{clone()} typically returns an object of the same class and "value"
as the input \code{x}.
}
\description{
Clone / duplicate an object
\code{clone()} physically duplicates objects
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
