% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster-statistics}
\alias{cluster-statistics}
\alias{SumOfRanges}
\alias{SumOfVariances}
\alias{SumOfVars}
\title{Cluster size statistics}
\usage{
SumOfRanges(mapping, cluster = 1)

SumOfVariances(mapping, cluster = 1)

SumOfVars(mapping, cluster = 1)
}
\arguments{
\item{mapping}{Matrix in which each row lists the coordinates of a point
in a Euclidian space.}

\item{cluster}{Optional integer vector specifying the cluster or group to
which each row in \code{mapping} belongs.}
}
\value{
\code{SumOfRanges()} returns a numeric specifying the sum of ranges
within each cluster across all dimensions.

\code{SumOfRanges()} returns a numeric specifying the sum of variances
within each cluster across all dimensions.
}
\description{
Cluster size statistics
}
\seealso{
Other tree space functions: 
\code{\link{MSTSegments}()},
\code{\link{MapTrees}()},
\code{\link{MappingQuality}()},
\code{\link{SpectralEigens}()},
\code{\link{median.multiPhylo}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree space functions}
