% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{ClusteringInfo}
\alias{ClusteringInfo}
\alias{ClusteringInfoSplits}
\title{Clustering information content of all partitions within a tree}
\usage{
ClusteringInfo(tree)

ClusteringInfoSplits(splits)
}
\arguments{
\item{tree}{A tree of class \code{phylo}, a list of trees, or a \code{multiPhylo} object.}
}
\value{
Information content, in bits.
}
\description{
Sums the clustering information content (Meila 2007) across each partition within a
phylogenetic tree.  This value will be greater than the total information
content of the tree where a tree contains multiple partitions, as
these partitions will contain mutual information.
}
\details{
Note that Meila (2007) and Vinh \emph{et al}. (2010) deal with entropy, which
denotes the bits necessary to denote the cluster to which each tip belongs,
i.e. bits/tip (see Vinh \emph{et al.} 2010).
This function multiplies the entropy by the number of tips to produce a
measure of the information content of the tree as a whole.
}
\section{Functions}{
\itemize{
\item \code{ClusteringInfoSplits}: Takes splits instead of trees
}}

\references{
\insertRef{Meila2007}{TreeDist}
}
\author{
Martin R. Smith
}
\keyword{internal}
