% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTable.R
\name{ClusterTable-methods}
\alias{ClusterTable-methods}
\alias{as.matrix.ClusterTable}
\alias{print.ClusterTable}
\alias{summary.ClusterTable}
\title{S3 methods for \code{ClusterTable} objects}
\usage{
\method{as.matrix}{ClusterTable}(x, ...)

\method{print}{ClusterTable}(x, ...)

\method{summary}{ClusterTable}(object, ...)
}
\arguments{
\item{x, object}{Object of class \code{ClusterTable}.}

\item{\dots}{Additional arguments for consistency with S3 methods.}
}
\description{
S3 methods for \code{\link{ClusterTable}} objects.
}
\examples{
clustab <- as.ClusterTable(TreeTools::BalancedTree(6))
as.matrix(clustab)
print(clustab)
summary(clustab)
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
