% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{CompareAll}
\alias{CompareAll}
\title{Distances between each pair of trees}
\usage{
CompareAll(x, Func, FUN.VALUE = Func(x[[1]], x[[1]]), ...)
}
\arguments{
\item{x}{List of trees, in the format expected by \code{Func}.}

\item{Func}{distance function returning distance between two trees,
e.g. \link[=path.dist]{phangorn::treedist}.}

\item{FUN.VALUE}{Format of output of \code{Func()}, passed to \code{\link{vapply}}.
Generated by running \code{Func(x[[1]], x[[1]])} if not specified.}
}
\value{
Distance matrix of class \code{dist} detailing distance between each pair
of trees.
Identical trees are assumed to have zero distance.
}
\description{
Distances between each pair of trees
}
\examples{
  library(TreeTools)
  trees <- list(BalancedTree(1:8), PectinateTree(1:8), PectinateTree(c(4:1, 5:8)))
  CompareAll(trees, NNIDist)
  
  # Or optionally, specifying a FUN.VALUE:
  dist <- CompareAll(trees, NNIDist, FUN.VALUE = vector('list', 3))
  as.matrix(dist)
  
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{pairwise tree distances}
