% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_transfer_dist_all_pairs}
\alias{cpp_transfer_dist_all_pairs}
\title{All-pairs transfer dissimilarity (OpenMP)}
\usage{
cpp_transfer_dist_all_pairs(splits_list, n_tip, scale, n_threads)
}
\arguments{
\item{splits_list}{List of raw matrices (one per tree).}

\item{n_tip}{Number of tips.}

\item{scale}{Logical: use scaled transfer dissimilarity?}

\item{n_threads}{Number of OpenMP threads.}
}
\value{
Numeric vector of length choose(N,2) in dist order.
}
\description{
All-pairs transfer dissimilarity (OpenMP)
}
\keyword{internal}
