% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_transfer_dist_cross_pairs}
\alias{cpp_transfer_dist_cross_pairs}
\title{Cross-pairs transfer dissimilarity (OpenMP)}
\usage{
cpp_transfer_dist_cross_pairs(splits_a, splits_b, n_tip, scale, n_threads)
}
\arguments{
\item{splits_a, splits_b}{Lists of raw matrices.}

\item{n_tip}{Number of tips.}

\item{scale}{Logical: use scaled transfer dissimilarity?}

\item{n_threads}{Number of OpenMP threads.}
}
\value{
Numeric matrix of dimension \code{nA} x \code{nB}.
}
\description{
Cross-pairs transfer dissimilarity (OpenMP)
}
\keyword{internal}
