% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{DescendantEdges}
\alias{DescendantEdges}
\alias{AllDescendantEdges}
\title{Descendant Edges}
\usage{
DescendantEdges(edge, parent, child, nEdge = length(parent))

AllDescendantEdges(parent, child, nEdge = length(parent))
}
\arguments{
\item{edge}{number of the edge whose child edges are required}

\item{parent}{the first column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 1]}

\item{child}{the second column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 2]}

\item{nEdge}{number of edges (calculated from length(parent) if not supplied)}
}
\value{
`DescendantEdges` returns a logical vector stating whether each edge in turn is a descendant of the specified edge
        (or the edge itself)

`AllDescendantEdges` returns a matrix of class logical, with row N specifying whether each edge is a descendant of edge N
        (or the edge itself)
}
\description{
Quickly identifies edges that are 'descended' from a particular edge in a tree
}
\section{Functions}{
\itemize{
\item \code{AllDescendantEdges}: Quickly identifies edges that are 'descended' from each edge in a tree
}}

