% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{DoDescendants}
\alias{DoDescendants}
\title{TITLE GOES HERE}
\usage{
DoDescendants(edge1, edge2, nTip = length(edge1)/2 + 1, node,
  just.tips = FALSE, just.internal = FALSE, include.ancestor = FALSE)
}
\arguments{
\item{edge1}{parent nodes: from tree$edge[,1]}

\item{edge2}{parent nodes: from tree$edge[,2]}

\item{node}{number of an internal node}

\item{just.tips}{(logical) should return value include all nodes or just tips?}
}
\value{
This function returns a vector containing descendant nodes in numerical order
}
\description{
\code{DoDescendants} does something useful
}
\examples{
warning(to_do <- TRUE)

}
\author{
Martin R. Smith
}
