% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplot.R
\name{.ThreeDPlotServer}
\alias{.ThreeDPlotServer}
\title{Helper function for shiny 3D plotting}
\usage{
.ThreeDPlotServer(
  input,
  output,
  distances,
  clusterings,
  mapping,
  mstEnds,
  mstSize,
  pointCols,
  thinnedTrees,
  StrainCol,
  PlotSize,
  mode3D
)
}
\description{
Helper function for shiny 3D plotting
}
\keyword{internal}
