% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{.TreeDistance}
\alias{.TreeDistance}
\title{Calculate distance between trees, or lists of trees}
\usage{
.TreeDistance(Func, tree1, tree2, checks = TRUE, ...)
}
\arguments{
\item{checks}{Logical specifying whether to perform basic sanity checks to
avoid crashes in C++.}
}
\description{
Calculate distance between trees, or lists of trees
}
\seealso{
\code{\link{CalculateTreeDistance}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
