% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoTreeSearch.R
\name{DoTreeSearch}
\alias{DoTreeSearch}
\title{DoTreeSearch}
\usage{
DoTreeSearch(tree, data, ParsimonyScorer = phangorn:::fitch, method = "NNI",
  maxiter = 100, maxhits = 20, forest.size = 1, cluster = NULL,
  track = 1, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with tip labels in the order generated by 
\code{\link{RenumberTips}}, i.e. corresponding corresponding the sequence of taxa 
in the corresponding Morphy object.}
}
\description{
Performs a tree search
}
\details{
Does the hard work of searching for a most parsimonious tree.
End-users are expected to access this function through its wrapper, TreeSearch
It is also called directly by Ratchet and Sectorial functions
}
\author{
Martin R. Smith
}
\keyword{internal}
