% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSearch.R
\name{DoTreeSearch}
\alias{DoTreeSearch}
\title{DoTreeSearch}
\usage{
DoTreeSearch(tree, data, TreeScorer = FitchScore, Rearrange = TBR,
  maxIter = 100, maxHits = 20, forestSize = 1, cluster = NULL,
  verbosity = 1, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}, with internal numbering in cladewise order
(use \code{\link{Preorder}(tree)} or (slower) \code{\link{Cladewise}(tree)})}

\item{data}{dataset in the format expected by TreeScorer}

\item{TreeScorer}{function to generate optimality score; defaults to \code{\link{FitchScore}}}

\item{Rearrange}{function to be used for tree rearrangements: probably \link{\code{NNI}},
\link{\code{SPR}} or \link{\code{TBR}}}

\item{maxIter}{Maximum iterations}

\item{maxHits}{stop search after finding optimal score \code{maxHits} times}

\item{forestSize}{number of trees to store in memory}

\item{verbosity}{Verbosity of reporting}
}
\value{
a tree of class \code{phylo} with attributes "hits" (number of times hit) and "pscore"
        (score given by TreeScorer)
}
\description{
Performs a tree search
}
\details{
Does the hard work of searching for a most parsimonious tree.
End-users are expected to access this function through its wrapper, TreeSearch
It is also called directly by Ratchet and Sectorial functions
}
\author{
Martin R. Smith
}
\keyword{internal}
