% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNI.R
\name{DoubleNNI}
\alias{DoubleNNI}
\title{Double NNI}
\usage{
DoubleNNI(parent, child, edgeToBreak)
}
\arguments{
\item{parent}{tree$edge[, 1]}

\item{child}{tree$edge[, 2]}

\item{edgeToBreak}{index of edge to break}
}
\value{
the \code{tree$edge} parameter of the two trees consistent with the specified rearrangement
}
\description{
Returns the edge parameter of the two trees consistent with the speficied \acronym{NNI} rearrangement
}
\author{
Martin R. Smith
}
\keyword{internal}
