% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{DropTip}
\alias{DropTip}
\title{TITLE GOES HERE}
\usage{
DropTip(phy, tip, trim.internal = TRUE, subtree = FALSE, root.edge = 0,
  rooted = is.rooted(phy), interactive = FALSE)
}
\arguments{
\item{PARAM}{is a parameter you should send to it}
}
\value{
This function returns :
}
\description{
\code{FUNCTIONNAME} does something useful
}
\examples{
to_do <- TRUE

}
\author{
Martin R. Smith
}
