% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{EdgeAncestry}
\alias{EdgeAncestry}
\title{EdgeAncestry}
\usage{
EdgeAncestry(edge, parent, child, stopAt = (parent == min(parent)))
}
\arguments{
\item{edge}{number of the edge whose child edges are required}

\item{parent}{the first column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 1]}

\item{child}{the second column of the edge matrix of a tree of class \code{\link{phylo}}, i.e. tree$edge[, 2]}

\item{stopAt}{number of the edge at which the search should terminate; defaults to the root edges}
}
\value{
a logical vector stating whether each edge in turn is a descendant of the specified edge
}
\description{
Descendant Edges
}
\details{
Quickly identifies edges that are 'ancestral' to a particular edge in a tree
}
\author{
Martin R. Smith
}
