% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{entropy_int}
\alias{entropy_int}
\title{Calculate entropy of integer vector of counts}
\usage{
entropy_int(n)
}
\arguments{
\item{n}{a vector of integer counts}
}
\value{
\code{entropy_int()} returns a numeric corresponding to the entropy of
each observation, in bits.
}
\description{
Wrapper for C++ function; no input checking is performed.
\code{\link[=Ntropy]{Ntropy()}} is better suited for use where performance is not critical.
}
\keyword{internal}
