% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{Fitch}
\alias{Fitch}
\alias{FitchScore}
\title{Fitch score}
\usage{
Fitch(tree, dataset, TipData = TipsAreNames, at = attributes(dataset),
  FitchFunction = C_Fitch_Score)

FitchScore(tree, dataset, TipData = NULL, at = attributes(dataset))
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{dataset}{A list or matrix whose list entries, rows or columns (as specified in TipData)
correspond to the character data associated with the tips of the tree. 
Likely of class \code{phyDat}.  The number of characters should be stored
as the attribute dataset$nr, and the weight of each character in dataset$weight;
if dataset is a matrix, they will be calculated automatically (weights set to 1).}

\item{TipData}{function to be used to match tree tips to dataset; probably one of 
\code{TipsAreNames}, \code{TipsAreRows}, or \code{TipsAreColumns}}

\item{at}{Attributes of the dataset (looked up automatically if not supplied)}

\item{FitchFunction}{function to be used to calculte parsimony score.}
}
\value{
A vector listing the number of 'parsimony steps' calculated for each character
}
\description{
Fitch score
}
\section{Functions}{
\itemize{
\item \code{FitchScore}: returns the parsimony score only
}}

