% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{GeneralizedRF}
\alias{GeneralizedRF}
\title{Generalized Robinson-Foulds distance}
\usage{
GeneralizedRF(splits1, splits2, PairScorer, maximize, reportMatching, ...)
}
\arguments{
\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{PairScorer}{function taking four arguments, \code{splits1}, \code{splits2},
\code{nSplits1}, \code{nSplits2}, which should return the score of each pair of splits
in a two-dimensional matrix.  Additional parameters may be specified via
\dots.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{\dots}{Additional parameters to \code{PairScorer}}
}
\value{
The results of \code{TreeDistanceReturn} under the parameters provided
}
\description{
Calculate Generalized Robinson-Foulds distance from splits.
}
\author{
Martin R. Smith
}
\keyword{internal}
