% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPart.R
\name{HPart}
\alias{HPart}
\alias{as.HPart}
\alias{as.HPart.HPart}
\title{Hierarchical partition structure}
\usage{
as.HPart(tree, tipLabels)

\method{as.HPart}{HPart}(tree, tipLabels = NULL)
}
\description{
Hierarchical partition structure
}
