% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_kendall-colijn.R
\name{KendallColijn}
\alias{KendallColijn}
\alias{KCVector}
\title{Kendall-Colijn distance}
\usage{
KendallColijn(tree1, tree2)

KCVector(tree)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree}{A tree of class \code{\link[ape:read.tree]{phylo}}.}
}
\description{
The Kendall-Colijn distance is related to the path difference.
}
\details{
It works by measuring, for each pair of tips, the distance from the most recent
common ancestor of those tips and the root node.  For a given tree, this
produces a vector of values recording the distance-from-the-root of each
most recent common ancestor of each pair of tips.

Two trees are compared by taking the Euclidian distance between the
respective vectors.  This is calculated by taking the square root of the sum
of the squares of the differences between the vectors.

This metric emphasizes the position of the root; the path difference
instead measures the distance of the last common ancestor of each pair
of tips from the tips themselves, i.e. the length of the path from one
tip to another.
}
\section{Functions}{
\itemize{
\item \code{KCVector}: Creates vectors that characterise a rooted tree
}}

\references{
\insertRef{Kendall2016}{TreeDist}
}
\seealso{
\href{https://cran.r-project.org/web/packages/treespace/vignettes/introduction.html}{treespace::treeDist},
a more sophisticated, if more cumbersome, implementation that supports
lambda > 0, i.e. use of edge lengths in tree comparison.

Other tree distances: \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{RobinsonFouldsInfo}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith
}
\concept{tree distances}
