% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanspp.R
\name{kmeanspp}
\alias{kmeanspp}
\title{k-means++ clustering}
\usage{
kmeanspp(x, k = 2, nstart = 10, ...)
}
\arguments{
\item{x}{Numeric matrix of data, or an object that can be coerced to such a
matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{k}{Integer specifying the number of clusters, \emph{k}.}

\item{nstart}{Positive integer specifying how many random sets should be
chosen}

\item{\dots}{additional arguments passed to \code{\link[stats]{kmeans}}}
}
\description{
k-means++ clustering \insertCite{Arthur2007}{TreeDist} improves the speed and
accuracy of standard \code{\link[stats]{kmeans}} clustering by selecting
random seeds that are far from existing cluster centroids.
}
\examples{
# Generate random points
set.seed(1)
n <- 100
x <- rbind(
  matrix(rnorm(n), ncol = 2),
  matrix(runif(n * 2, -1, 1), ncol = 2)
)

# Compute clustering
clusters <- kmeanspp(x, k = 5)

# Plot
plot(x, col = clusters$cluster, pch = 16)
}
\references{
\insertAllCited{}#'
}
\seealso{
\code{\link[stats]{kmeans}}
}
\author{
Modified by Martin R. Smith from \href{https://rdrr.io/cran/LICORS/src/R/kmeanspp.R}{code} by Georg M. Goerg.
}
