% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lap.R
\name{LAPJV}
\alias{LAPJV}
\title{Solve linear assignment problem using LAPJV}
\usage{
LAPJV(x)
}
\arguments{
\item{x}{Matrix to solve.}
}
\value{
A list with two entries: \code{score}, the score of the optimal matching;
and \code{matching}, the columns matched to each row of the matrix in turn.
}
\description{
Solve the Linear Assignment Problem with the algorithm of
Jonker & Volgenant (1987).
}
\details{
The Linear Assignment Problem seeks to match each row of a matrix with a
column, such that the cost of the matching is minimized.

The JV approach improves on the Hungarian algorithm, which is implemented
in \code{\link[clue:solve_LSAP]{clue::solve_LSAP}}.
}
\references{
"A Shortest Augmenting Path Algorithm for Dense and Sparse Linear
Assignment Problems," Computing 38, 325-340, 1987

R. Jonker and A. Volgenant, University of Amsterdam.
}
\author{
C++ code by Roy Jonker @ MagicLogic Optimization Inc.
e-mail: roy_jonker@magiclogic.com

Changed 2016-05-13 by Yong Yang(yongyanglink@gmail.com) in column reduction
part according to matlab version of LAPJV algorithm (Yi Cao)
https://github.com/yongyanghz/LAPJV-algorithm-c/blob/master/LAPJV/lap.cpp
}
