% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ListToMatrix}
\alias{ListToMatrix}
\title{Edge list to edge matrix}
\usage{
ListToMatrix(edgeList)
}
\arguments{
\item{edgeList}{tree edges in the format list(parent, child).}
}
\value{
edges in the format expected by \code{tree$edge},
where \code{tree} is a tree of class \code{phylo}.
}
\description{
Edge list to edge matrix
}
\keyword{internal}
