% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_mast.R
\name{MASTSize}
\alias{MASTSize}
\title{Size of maximum agreement subtree}
\usage{
MASTSize(tree1, tree2, tree = FALSE, rooted = TRUE)
}
\arguments{
\item{tree1, tree2}{Single trees of class \code{phylo} to undergo comparison.}

\item{tree}{Logical specifying whether to return the maximum agreement
subtree; if \code{FALSE}, only its size will be returned.}

\item{rooted}{Logical specifying whether to treat the trees as rooted.}
}
\value{
An integer specifying the size of the maximum agreement subtree.
}
\description{
Calculate the maximum agreement subtree between two phylogenetic trees.
}
\references{
\insertref{Valiente2009}{TreeDist}
}
\seealso{
\code{\link[phangorn:mast]{phangorn::mast}}, a slower, all-R implementation that also returns
the tips contained within the subtree.

Other tree distances: \code{\link{JaccardRobinsonFoulds}},
  \code{\link{KendallColijn}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{RobinsonFouldsInfo}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
