% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_msd.R
\name{MatchingSplitDistance}
\alias{MatchingSplitDistance}
\alias{MatchingSplitDistanceSplits}
\title{Matching Split Distance}
\usage{
MatchingSplitDistance(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

MatchingSplitDistanceSplits(splits1, splits2, nTip = attr(splits1,
  "nTip"), normalize = TRUE, reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{nTip}{Integer specifying the number of tips in each split.}
}
\value{
An array of numerics providing the distances between each pair of
trees in \code{tree1} and \code{tree2}.
}
\description{
Implements the Matching Split Distance for unrooted binary phylogenetic
trees of Bogdanowicz and Giaro (2012).
}
\section{Functions}{
\itemize{
\item \code{MatchingSplitDistanceSplits}: Calculate Matching Split Distance from splits instead of trees.
}}

\section{Normalization}{


A normalization value or function must be provided in order to return a
normalized value.  If you are aware of a generalised formula, please
let me know by
\href{https://github.com/ms609/TreeDist/issues/new}{creating a GitHub issue}
so that it can be implemented.
}

\references{
\insertRef{Bogdanowicz2012}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{JaccardRobinsonFoulds}},
  \code{\link{KendallColijn}}, \code{\link{MASTSize}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{RobinsonFouldsInfo}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
