% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{MCITree}
\alias{MCITree}
\alias{GreedyCI}
\title{Maximum Clade Information Tree}
\usage{
MCITree(trees, info = "phylogenetic", check.tips = TRUE)

GreedyCI(trees, info = "phylogenetic", check.tips = TRUE)
}
\arguments{
\item{trees}{List of \code{phylo} objects, optionally with class \code{multiPhylo}.}

\item{info}{Abbreviation of "phylogenetic" or "clustering", specifying
the concept of information to employ.}

\item{check.tips}{Logical specifying whether to renumber leaves such that
leaf numbering is consistent in all trees.}
}
\value{
\code{MCITree()} returns the binary tree with the highest information
content, given the posterior distribution
}
\description{
Analogous to the Maximum Clade Credibility tree:
Selects the tree from the posterior distribution whose clades have the
highest information content.
Generate the MCC tree by specifying \code{info = "credibility"}.
}
\examples{
library("TreeTools", quietly = TRUE)
trees <- as.phylo(24:40, 16)

# Maximum Clade Information tree
mci <- MCITree(trees)
SplitwiseInfo(mci)
plot(mci)
p <- SplitFrequency(mci, trees) / length(trees)
LabelSplits(mci, round(p * 100), "\%", bg = SupportColor(p))

# Compare with Maximum Clade Credibility tree
mcc <- MCITree(trees, "credibility")
plot(mcc)
p <- SplitFrequency(mcc, trees) / length(trees)
LabelSplits(mcc, round(p * 100), "\%", bg = SupportColor(p))
SplitwiseInfo(mcc)

}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
