% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{MeilaVariationOfInformation}
\alias{MeilaVariationOfInformation}
\alias{MeilaMutualInformation}
\title{Use Variation of Information to compare splits as clusterings}
\usage{
MeilaVariationOfInformation(split1, split2)

MeilaMutualInformation(split1, split2)
}
\arguments{
\item{split1, split2}{Logical vectors listing leavess in a consistent order,
identifying each leaf as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the split in question.}
}
\value{
Variation of information, measured in bits.

Mutual information, measured in bits.
}
\description{
Compare a pair of splits solely as clusterings of taxa, disregarding their
phylogenetic information, using the variation of clustering information
proposed by Meila (2007).
}
\details{
This is equivalent to the mutual clustering information (Vinh \emph{et al.} 2010).

Whereas Meila and Vinh \emph{et al}. are concerned with the entropy, i.e. the
number of bits required to encode each tip, we are interested in the
total information content, i.e. the entropy multiplied by the number of tips.
This is the value that the function returns.
}
\section{Functions}{
\itemize{
\item \code{MeilaMutualInformation}: Mutual clustering information of two splits
}}

\examples{
  # Maximum variation = information content of each split separately
  A <- TRUE
  B <- FALSE
  MeilaVariationOfInformation(c(A,A,A,B,B,B), c(A,A,A,A,A,A))
  Entropy(c(3, 3) / 6) + Entropy(c(0, 6) / 6)
  
  # Minimum variation = 0
  MeilaVariationOfInformation(c(A,A,A,B,B,B), c(A,A,A,B,B,B))
  
  # Not always possible for two evenly-sized splits to reach maximum
  # variation of information
  Entropy(c(3, 3) / 6) * 2  # = 2
  MeilaVariationOfInformation(c(A,A,A,B,B,B), c(A,B,A,B,A,B)) # < 2
  
  # Phylogenetically uninformative groupings contain spliting information
  Entropy(c(1, 5) / 6)
  MeilaVariationOfInformation(c(B,A,A,A,A,A), c(A,A,A,A,A,B))


}
\references{
{
\insertRef{Meila2007}{TreeDist}
\insertRef{Vinh2010}{TreeDist}
}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
