% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpl_tree_search.R
\name{MorphyBootstrap}
\alias{MorphyBootstrap}
\title{Bootstrap tree search with inapplicable data}
\usage{
MorphyBootstrap(edgeList, morphyObj, maxIter, maxHits, verbosity = 1L, ...)
}
\arguments{
\item{edgeList}{a list containing the following:
- vector of integers corresponding to the parent of each edge in turn
- vector of integers corresponding to the child of each edge in turn
- (optionally) score of the tree
- (optionally, if score provided) number of times this score has been hit}

\item{morphyObj}{A morphy object, perhaps created with \code{\link{PhyDat2Morphy}}.}

\item{maxIter}{maximum number of iterations to perform in tree search}

\item{maxHits}{maximum number of hits to accomplish in tree search}

\item{verbosity}{Level of detail to display in console: larger numbers provide
more verbose feedback to the user.}

\item{\dots}{further parameters to send to \code{MorphyTreeSearch}}
}
\value{
A tree that is optimal under a random sampling of the original characters
}
\description{
Bootstrap tree search with inapplicable data
}
