% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{VariationOfPartitionInfo}
\alias{VariationOfPartitionInfo}
\alias{ExpectedVariation}
\alias{MutualPartitionInfo}
\alias{MutualPartitionInfoSplits}
\alias{MutualClusteringInfoSplits}
\title{Mutual Partition Information}
\usage{
VariationOfPartitionInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

ExpectedVariation(tree1, tree2, samples = 1000)

MutualPartitionInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

MutualPartitionInfoSplits(splits1, splits2, reportMatching = FALSE)

MutualClusteringInfoSplits(splits1, splits2, normalize = TRUE,
  reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in 'details' below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{samples}{Integer specifying how many samplings to obtain;
accuracy of estimate increases with \code{sqrt(samples)}.}

\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}
}
\description{
#TODO document!
}
\section{Functions}{
\itemize{
\item \code{VariationOfPartitionInfo}: Variation of partition information between two trees

\item \code{ExpectedVariation}: Estimate expected Variation of Information
and Mutual Information for a pair of trees of a given topology.

\item \code{MutualPartitionInfoSplits}: Takes splits instead of trees

\item \code{MutualClusteringInfoSplits}: Takes splits instead of trees
}}

\references{
\insertRef{SmithDist}{TreeSearch}
}
\seealso{
Other Tree distance: \code{\link{MatchingSplitDistance}},
  \code{\link{MutualArborealInfo}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{VariationOfClusteringInfo}}
}
\author{
Martin R. Smith
}
\concept{Tree distance}
