% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_info_extra_step.r
\name{N1Spr}
\alias{N1Spr}
\alias{IC1Spr}
\title{Number of trees on SPR step away
Formula given by Given by Allen and Steel 2001.}
\usage{
N1Spr(n)

IC1Spr(n)
}
\arguments{
\item{n}{Number of tips in tree.}
}
\description{
Number of trees on SPR step away
Formula given by Given by Allen and Steel 2001.
}
\section{Functions}{
\itemize{
\item \code{IC1Spr}: Information content of trees 0 or 1 SPR step from tree with n tips.
}}

\references{
\insertRef{Allen2001}{TreeSearch}
}
