% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNI.R
\name{NNI}
\alias{NNI}
\title{Tree rearrangement functions}
\usage{
NNI(tree, edgeToBreak = NULL)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{edgeToBreak}{(optional) integer specifying the index of an edge to bisect/prune, generated randomly if not specified}

\item{mergeEdges}{(optional) vector of length 1 or 2, listing edge(s) to be joined:
In SPR, this is where the pruned subtree will be reconnected.
In TBR, these edges will be reconnected (so must be on opposite
sides of \code{edgeToBreak}); if only a single edge is specified,
the second will be chosen at random}
}
\value{
Returns a tree with class \code{phylo}.
}
\description{
These functions performs a single random \acronym{TBR}, \acronym{SPR} or \acronym{NNI} iteration.
}
\details{
Performs a single iteration of the nearest-neigbour interchange, subtree pruning and regrafting,
or tree bisection and reconnection algorithms.
NNI and SPR are based on the corresponding phangorn functions, but have been re-coded to 
improve their speed.

Branch lengths are not supported.
}
\examples{
library(ape)
tree <- ape:::rtree(20, br=NULL)
NNI(tree)
SPR(tree)
TBR(tree)

}
\references{
The algorithms are summarized in
Felsenstein, J. 2004. \cite{Inferring Phylogenies.} Sinauer Associates, Sunderland, Massachusetts.
}
\author{
Martin R. Smith
}
