% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_tree.R
\name{NNI}
\alias{NNI}
\title{Tree rearrangement functions

These functions performs a single random \acronym{TBR}, \acronym{SPR} or \acronym{NNI} iteration.

Performs a single iteration of the nearest-neigbour interchange, subtree pruning and regrafting,
or tree bisection and reconnection algorithms.
NNI and SPR are based on the corresponding phangorn functions, but have been re-coded to 
improve their speed.

Branch lengths are not supported.}
\usage{
NNI(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{edge.to.break}{the index of an edge to bisect, generated randomly if not specified}
}
\value{
Returns a tree with class \code{phylo}.
}
\examples{
tree <- ape:::rtree(20, br=NULL)
NNI(tree)
SPR(tree)
TBR(tree)
}
\references{
The algorithms are summarized in
Felsenstein, J. 2004. \cite{Inferring Phylogenies.} Sinauer Associates, Sunderland, Massachusetts.
}
\author{
Martin R. Smith
}
