% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_nni.R
\name{NNIDist}
\alias{NNIDist}
\title{Approximate the NNI distance}
\usage{
NNIDist(tree1, tree2)
}
\arguments{
\item{tree1, tree2}{Individual trees of class \code{phylo}}
}
\description{
An approximation of the NNI distance based on Li et al. 1996.
}
\examples{
  library('TreeTools')
  tree1 <- BalancedTree(7)
  tree2 <- PectinateTree(7)
  
  NNIDist(tree1, tree2)

}
\references{
\insertRef{Li1996}{TreeDist}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
