% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{NormalizeInfo}
\alias{NormalizeInfo}
\title{Normalize information against total present in both starting trees}
\usage{
NormalizeInfo(unnormalized, tree1, tree2, InfoInTree, infoInBoth = NULL,
  how = TRUE, Combine = "+", ...)
}
\arguments{
\item{unnormalized}{Numeric value to be normalized.}

\item{tree1, tree2}{Trees from which \code{unnormalized} was calculated}

\item{InfoInTree}{Function to calculate the information content of each tree}

\item{infoInBoth}{Numeric speecifying information content of both trees
independently (optional)}

\item{how}{Method for normalization}

\item{\dots}{Additional parameters to \code{InfoInTree`` or }how`.}

\item{Func}{Function that takes as inputs \code{tree1Info} and \code{tree2Info}, and
returns a normalizing constant against which to divide \code{unnormalized}.}
}
\description{
Normalize information against total present in both starting trees
}
\author{
Martin R. Smith
}
\keyword{internal}
