% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R
\name{NyeTreeSimilarity}
\alias{NyeTreeSimilarity}
\title{Nye et al. (2006) tree comparison}
\usage{
NyeTreeSimilarity(tree1, tree2, reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}
}
\description{
Implements the tree comparison metric of Nye \emph{et al}. (2006).
}
\references{
\insertRef{Nye2006}{TreeSearch}
}
\author{
Martin R. Smith
}
\concept{Tree distance}
