% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_nye.R
\name{NyeTreeSimilarity}
\alias{NyeTreeSimilarity}
\alias{NyeSplitSimilarity}
\title{Nye \emph{et al}. (2006) tree comparison}
\usage{
NyeTreeSimilarity(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

NyeSplitSimilarity(splits1, splits2, normalize = TRUE,
  reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}
}
\description{
Implements the tree comparison metric of Nye \emph{et al}. (2006).
In short, this finds the optimal matching that pairs each branch from
one tree with a branch in the second, where matchings are scored according to
the size of the largest bipartition that is consistent with both of them,
normalized against the Jaccard index.
}
\section{Functions}{
\itemize{
\item \code{NyeSplitSimilarity}: Takes splits instead of trees
}}

\section{Normalization}{


If \code{normalize = TRUE}, then results will be rescaled from zero to a maximum
value calculated by #TODO detail
}

\references{
\insertRef{Nye2006}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{KendallColijn}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{MutualArborealInfo}},
  \code{\link{MutualClusteringInfo}}
}
\author{
Martin R. Smith
}
\concept{tree distances}
