% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_nye.R
\name{NyeTreeSimilarity}
\alias{NyeTreeSimilarity}
\alias{NyeSplitSimilarity}
\title{Nye \emph{et al}. (2006) tree comparison}
\usage{
NyeTreeSimilarity(tree1, tree2 = tree1, similarity = TRUE,
  normalize = FALSE, normalizeMax = TRUE, reportMatching = FALSE)

NyeSplitSimilarity(splits1, splits2, nTip = attr(splits1, "nTip"),
  reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{similarity}{Logical specifying whether to report the result as a tree
similarity, rather than a difference.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{normalizeMax}{When calculating similarity, normalize against the
maximum number of partitions that could have been present (\code{TRUE}),
or the number of partitions that were actually observed (\code{FALSE})?
Defaults to the number of partitions in the better-resolved tree; set
\code{normalize = pmin.int} to use the number of partitions in the less resolved
tree.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{splits1}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}

\item{nTip}{Integer specifying the number of tips in each split.}
}
\value{
An array of numerics providing the distances between each pair of
trees in \code{tree1} and \code{tree2}.
}
\description{
Implements the tree comparison metric of Nye \emph{et al}. (2006).
In short, this finds the optimal matching that pairs each branch from
one tree with a branch in the second, where matchings are scored according to
the size of the largest bipartition that is consistent with both of them,
normalized against the Jaccard index.  This is equivalent to (but
slightly faster than) \code{\link{JaccardRobinsonFoulds}
(tree1, tree2, k = 1, arboreal = FALSE)}.
}
\details{
The measure is defined as a similarity score.  If \code{similarity = FALSE}, the
similarity score will be converted into a distance by doubling it and
subtracting it from the number of partitions present in both trees.
This ensures consistency with \code{JaccardRobinsonFoulds}.
}
\section{Functions}{
\itemize{
\item \code{NyeSplitSimilarity}: Calculate tree similarity from splits
instead of trees.
}}

\section{Normalization}{


If \code{normalize = TRUE} and \code{similarity = TRUE}, then results will be rescaled
from zero to one by dividing by the maximum value possible for any pair
of trees with  \emph{n} terminals, $n - 3$.
You may wish to normalize instead against the number of partitions present
in the smaller tree, which represents the maximum value possible for a pair
of trees with the specified topologies (\code{normalize = pmin.int}), or the
number of partitions in the most resolved tree (\code{normalize = pmax.int}).

If \code{normalize = TRUE} and \code{similarity = FALSE}, then results will be rescaled
from zero to one by dividing by the total number of partitions in the pair
of trees being considered.
}

\examples{
library('TreeTools')
NyeTreeSimilarity(BalancedTree(8), PectinateTree(8))
NyeTreeSimilarity(lapply(0:5, as.phylo, nTip = 8), PectinateTree(8))
NyeTreeSimilarity(lapply(0:5, as.phylo, nTip = 8), similarity = FALSE)

}
\references{
\insertRef{Nye2006}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{JaccardRobinsonFoulds}},
  \code{\link{KendallColijn}}, \code{\link{MASTSize}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{RobinsonFouldsInfo}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
