% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneOverlap}
\alias{oneOverlap}
\title{One Overlap}
\format{An object of class \code{list} of length 500.}
\usage{
oneOverlap
}
\description{
Information content of a single overlap
Internal cache, called from within \code{MutualPhylogeneticInfoSplits}.
}
\keyword{datasets}
