% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{RandomTree}
\alias{RandomTree}
\title{Generate random tree topology from dataset}
\usage{
RandomTree(dataset)
}
\arguments{
\item{dataset}{A dataset in \code{\link[phangorn]{phyDat}} format}
}
\description{
Generate random tree topology from dataset
}
\author{
Martin R. Smith
}
