% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomTreeScore.R
\name{RandomTreeScore}
\alias{RandomTreeScore}
\title{Parsimony score of random postorder tree}
\usage{
RandomTreeScore(nTip, morphyObj)
}
\arguments{
\item{nTip}{number of tips (minimum 3)}

\item{morphyObj}{A morphy object created with \code{\link{LoadMorphy}}.}
}
\value{
the parsimony score of a random tree, for the given Morphy object.
}
\description{
Parsimony score of random postorder tree
}
