% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratchet.R
\name{Ratchet}
\alias{Ratchet}
\title{Parsimony Ratchet}
\usage{
Ratchet(tree, data, TreeScorer = FitchScore, all = FALSE, outgroup = NULL,
  pratchiter = 100, searchiter = 5000, searchhits = 40, pratchhits = 10,
  track = 0, rearrangements = "NNI", suboptimal = 1e-08, ...)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}}

\item{data}{a dataset in the format required by TreeScorer}

\item{all}{Set to \code{TRUE} to report all MPTs encountered during the search, perhaps to analyze consensus}

\item{outgroup}{a vector specifying all tips in the outgroup; if unspecified then identical trees with different roots will be considered unique;}

\item{rearrangements}{method(s) to use when rearranging trees: 
a vector containing any combination of the strings "NNI", "SPR" or "TBR";}

\item{\dots}{other arguments to pass to subsequent functions.}

\item{concavity}{concavity constant for implied weighting (not currently implemented!)}

\item{maxit}{maximum ratchet iterations to perform;}

\item{maxIter}{maximum rearrangements to perform on each bootstrap or ratchet iteration;}

\item{maxHits}{maximum times to hit best score before terminating a tree search within a pratchet iteration;}

\item{k}{stop when k ratchet iterations have found the same best score;}

\item{verbosity}{larger numbers provides more verbose feedback to the user;}
}
\value{
This function returns a tree modified by parsimony ratchet iteration, retaining the position of the root.
}
\description{
\code{Ratchet} uses the parsimony ratchet (Nixon 1999) to search for a more parsimonious tree.
}
\examples{
Ratchet(RandomTree(Lobo.phy), SigSut.phy, outgroup='Cricocosmia')

}
\references{
Nixon, K. C. (1999). \cite{The Parsimony Ratchet, a new method for rapid parsimony analysis.} Cladistics, 15(4), 407-414. doi:\href{http://dx.doi.org/10.1111/j.1096-0031.1999.tb00277.x}{10.1111/j.1096-0031.1999.tb00277.x}
}
\seealso{
\code{\link{Ratchet}}

\code{\link{TreeSearch}}

\code{\link{SectorialSearch}}
}
\author{
Martin R. Smith

Adapted from \code{\link[phangorn]{pratchet}} in the \pkg{phangorn} package, which does not preserve the position of the root.
}
\keyword{tree}
