% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{RenumberTree}
\alias{RenumberTree}
\title{Renumber tree
Order edges and number nodes
Wrapper for the C function RENUMBER_TREE}
\usage{
RenumberTree(parent, child, nEdge = length(parent))
}
\value{
an edge matrix for a tree following the usual convention for edge and node numbering
}
\description{
Renumber tree
Order edges and number nodes
Wrapper for the C function RENUMBER_TREE
}
\keyword{internal}
