% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{RenumberTree}
\alias{RenumberTree}
\alias{RenumberEdges}
\title{Renumber a tree
Order edges and number nodes
Wrapper for the C function RENUMBER_TREE}
\usage{
RenumberTree(parent, child, nEdge = length(parent))

RenumberEdges(parent, child, nEdge = length(parent))
}
\value{
an edge matrix for a tree in following the usual preorder convention for edge and node numbering
}
\description{
Renumber a tree
Order edges and number nodes
Wrapper for the C function RENUMBER_TREE
}
\section{Functions}{
\itemize{
\item \code{RenumberEdges}: Instead returns a list containing two items corresponding to the new parent and child vectors
}}

\keyword{internal}
