% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_rf.R
\name{RobinsonFouldsInfo}
\alias{RobinsonFouldsInfo}
\alias{RobinsonFouldsInfoSplits}
\alias{RobinsonFoulds}
\alias{RobinsonFouldsMatching}
\alias{RobinsonFouldsSplits}
\title{(Information-adjusted) Robinson-Foulds distance}
\usage{
RobinsonFouldsInfo(tree1, tree2 = tree1, similarity = FALSE,
  normalize = FALSE, reportMatching = FALSE)

RobinsonFouldsInfoSplits(splits1, splits2, nTip = attr(splits1, "nTip"),
  reportMatching = FALSE)

RobinsonFoulds(tree1, tree2 = tree1, similarity = FALSE,
  normalize = FALSE, reportMatching = FALSE)

RobinsonFouldsMatching(tree1, tree2 = tree1, similarity = FALSE,
  normalize = FALSE, ...)

RobinsonFouldsSplits(splits1, splits2, nTip = attr(splits1, "nTip"),
  reportMatching = FALSE)
}
\arguments{
\item{tree1}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{similarity}{Logical specifying whether to report the result as a tree
similarity, rather than a difference.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{splits1}{Logical matrices where each row corresponds to a leaf,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a split, such that each leaf is identified as
a member of the ingroup (\code{TRUE}) or outgroup (\code{FALSE}) of the respective
split.}

\item{splits2}{Logical matrices where each row corresponds to a leaf,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a split, such that each leaf is identified as
a member of the ingroup (\code{TRUE}) or outgroup (\code{FALSE}) of the respective
split.}

\item{nTip}{Integer specifying the number of tips in each split.}

\item{\dots}{Not used.}
}
\value{
An array of numerics providing the distances between each pair of
trees in \code{tree1} and \code{tree2}.
}
\description{
Calculates the Robinson-Foulds distance, or the equivalent similarity
measure, optionally annotating matched splits and weighting splits
according to their phylogenetic information content.
}
\details{
\code{RobinsonFoulds} is an inefficient implementation of the Robinson-Foulds
distance, included for use with \code{\link{VisualizeMatching}}.
To generate the RF distance efficiently,
use the function \code{\link{ape}{treedist}}.

Note that if \code{reportMatching = TRUE}, the \code{pairScores} attribute returns
a logical matrix specifying whether each pair of splits is identical.

\code{RobinsonFouldsInfo} calculates the tree similarity or distance by summing
the phylogenetic information content of all splits that are (or are not)
identical in both trees.  Consequently, splits that are more likely
to be identical by chance alone make a smaller contribution to overall
tree distance, because their similarity is less remarkable.
}
\section{Functions}{
\itemize{
\item \code{RobinsonFouldsInfoSplits}: Calculate information-adjusted Robinson-Foulds
distance from splits instead of trees.

\item \code{RobinsonFoulds}: Robinson-Foulds distance, with option to
report matched splits.

\item \code{RobinsonFouldsMatching}: Matched splits, intended for use with
\code{\link{VisualizeMatching}}.

\item \code{RobinsonFouldsSplits}: Calculate Robinson-Foulds distance from splits
instead of trees.
}}

\section{Normalization}{

\itemize{
\item \code{RobinsonFoulds} is normalized against the total number of splits that
are present.
\item \code{RobinsonFouldsInfo} is normalized against the sum of the phylogenetic
information of all splits in both trees, treated independently.
}
}

\references{
\insertRef{Robinson1981}{TreeDist}
}
\seealso{
Other tree distances: \code{\link{JaccardRobinsonFoulds}},
  \code{\link{KendallColijn}}, \code{\link{MASTSize}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}},
  \code{\link{TreeDistance}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
