% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{SplitsRepeated}
\alias{SplitsRepeated}
\title{Are Splits Repeated?}
\usage{
SplitsRepeated(original, matches)
}
\arguments{
\item{original, matches}{Logical matrix describing the bipartition splits
defined by a tree (see [Quartet:Tree2Splits] for format).  The first row
of `original` must be `FALSE` for all splits.}
}
\value{
`SplitsRepeated` returns a logical vector of length `ncol(original)`,
specifying whether each split listed in `original` also appears in `matches`.
}
\description{
Determines whether any of the splits described in `original` are present in
`matches`.
}
\author{
Martin R. Smith
}
\keyword{internal}
