% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{SplitwiseInfo}
\alias{SplitwiseInfo}
\alias{ConsensusInfo}
\title{Information content of splits within a tree}
\usage{
SplitwiseInfo(x, p = NULL, sum = TRUE)

ConsensusInfo(trees, info = "phylogenetic", check.tips = TRUE)
}
\arguments{
\item{x}{A tree of class \code{phylo}, a list of trees, or a \code{multiPhylo} object.}

\item{p}{A vector of probabilities corresponding t oeach split in \code{x}.
Specify \code{TRUE} to calculate this vector using the node labels of each tree.}

\item{sum}{Logical: if \code{TRUE}, sum the information content of each split to
provide the total splitwise information content of the tree.}

\item{trees}{List of \code{phylo} objects, optionally with class \code{multiPhylo}.}

\item{info}{Abbreviation of 'phylogenetic' or 'clustering', specifying
the concept of information to employ. See \insertCite{Smith2020}{TreeDist}
or \href{https://ms609.github.io/TreeDist/articles/information.html}{vignette}
for details.}

\item{check.tips}{Logical specifying whether to renumber leaves such that
leaf numbering is consistent in all trees.}
}
\value{
\code{SplitwiseInfo()} returns the splitwise information content of the
tree (or of each split in turn, if \code{sum = FALSE}), in bits.

\code{ConsensusInfo()} returns the splitwise information content of the
majority rule consensus of \code{trees}.
}
\description{
Sum the phylogenetic information content for all splits within a
phylogenetic tree.  This value will be greater than the total information
content of the tree where a tree contains multiple splits, as
these splits will contain mutual information.
}
\examples{
SplitwiseInfo(TreeTools::PectinateTree(8))
tree <- ape::read.tree(text = "(a, b, (c, (d, e, (f, g)0.8))0.9);")
SplitwiseInfo(tree)
SplitwiseInfo(tree, TRUE)
library("TreeTools")
set.seed(0)
trees <- list(RandomTree(8), RootTree(BalancedTree(8), 1), PectinateTree(8))
cons <- consensus(trees, p = 0.5)
p <- SplitFrequency(cons, trees) / length(trees)
plot(cons)
LabelSplits(cons, signif(SplitwiseInfo(cons, p, sum = FALSE), 4))
ConsensusInfo(trees)
LabelSplits(cons, signif(ClusteringInfo(cons, p, sum = FALSE), 4))
ConsensusInfo(trees, 'clustering')
}
\seealso{
An introduction to the phylogenetic information content of a split is given
in \href{https://ms609.github.io/TreeTools/reference/SplitInformation.html}{\code{SplitInformation()}}
and in a \href{https://ms609.github.io/TreeDist/articles/information.html}{package vignette}.

Other information functions: 
\code{\link{ClusteringEntropy}()},
\code{\link{SplitEntropy}()},
\code{\link{SplitSharedInformation}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{information functions}
