% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_spr.R
\name{SPRDist}
\alias{SPRDist}
\title{Approximate Subtree Prune and Regraft distance}
\usage{
SPRDist(tree1, tree2 = NULL)
}
\arguments{
\item{tree1, tree2}{Single trees of class \code{phylo} to undergo comparison.}
}
\value{
\code{SPRDist()} returns a vector or distance matrix of distances
between trees.
}
\description{
Approximate the Subtree Prune and Regraft (SPR) distance.
}
\details{
This function is a wrapper for the function
\code{\link[phangorn:treedist]{SPR.dist()}} in the phangorn package.
It pre-processes trees to ensure that their internal representation does
not cause the \code{SPR.dist()} function to crash R.
}
\examples{
library('TreeTools')

SPRDist(BalancedTree(7), PectinateTree(7))

SPRDist(BalancedTree(7), as.phylo(0:2, 7))
SPRDist(as.phylo(0:2, 7), PectinateTree(7))

SPRDist(list(bal = BalancedTree(7), pec = PectinateTree(7)),
        as.phylo(0:2, 7))

CompareAll(as.phylo(30:33, 8), SPRDist)
  
}
\seealso{
Other tree distances: 
\code{\link{JaccardRobinsonFoulds}()},
\code{\link{KendallColijn}()},
\code{\link{MASTSize}()},
\code{\link{MatchingSplitDistance}()},
\code{\link{NNIDist}()},
\code{\link{NyeTreeSimilarity}()},
\code{\link{PathDist}()},
\code{\link{Robinson-Foulds}},
\code{\link{TreeDistance}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
