% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBR.R
\name{TBR}
\alias{TBR}
\alias{RootedTBR}
\title{TBR}
\usage{
TBR(tree, edgeToBreak = NULL, mergeEdges = NULL)

RootedTBR(tree, edgeToBreak = NULL, mergeEdges = NULL)
}
\arguments{
\item{tree}{A bifurcating tree of class \code{\link{phylo}}, with all nodes resolved;}

\item{edgeToBreak}{the index of an edge to bisect, generated randomly if not specified.}
}
\value{
This function returns a tree in \code{phyDat} format that has undergone one \acronym{TBR} iteration.
}
\description{
Tree bisection and reconnection

Rooted TBR
}
\details{
\code{TBR} performs a single random \acronym{TBR} iteration.

Branch lengths are not (yet) supported.
}
\section{Functions}{
\itemize{
\item \code{RootedTBR}: Perform \acronym{TBR} rearrangement, retaining position of root
}}

\examples{
{
library('ape')
tree <- rtree(20, br=NULL)
TBR(tree)
}
}
\references{
The \acronym{TBR} algorithm is summarized in
Felsenstein, J. 2004. \cite{Inferring Phylogenies.} Sinauer Associates, Sunderland, Massachusetts.
}
\seealso{
RootedTBR useful when the position of the root node should be retained.
}
\author{
Martin R. Smith
}
