% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPR.R, R/TBR.R
\name{SPRWarning}
\alias{SPRWarning}
\alias{TBRWarning}
\title{TBR Warning
Print a warning and return given tree}
\usage{
SPRWarning(tree, error)

TBRWarning(tree, error)
}
\arguments{
\item{tree}{tree to return}

\item{error}{error message to report}
}
\value{
the tree specified in tree
}
\description{
TBR Warning
Print a warning and return given tree
}
\section{Functions}{
\itemize{
\item \code{SPRWarning}: for SPR rearrangements
}}

\examples{
testFunction <- function (tree) {
 return(TBRWarning(tree, 'Message text'))
}
\dontrun{testFunction(0) # will trigger warning}


}
\author{
Martin R. Smith
}
\keyword{internal}
