% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{TBRWarning}
\alias{TBRWarning}
\title{TBR Warning
Print a warning and return given tree}
\usage{
return(TBRWarning(tree, 'Message text')
}
\arguments{
\item{tree}{tree to return}

\item{error}{error message to report}
}
\description{
TBR Warning
Print a warning and return given tree
}
\author{
Martin R. Smith
}
\keyword{internal}
