% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{TipsAreNames}
\alias{TipsAreNames}
\alias{TipsAreRows}
\alias{TipsAreColumns}
\title{Extract character data from dataset}
\usage{
TipsAreNames(data, tips)

TipsAreRows(data, tips)

TipsAreColumns(data, tips)
}
\arguments{
\item{data}{a matrix or list containing symbols associated with each tip; for example,
a dataset of class \code{phyDat}}

\item{tips}{vector detailing the tips to be selected, whether as their
names or numbers corresponding to their rows/columns}
}
\value{
an integer vector, listing the tokens associated with each character for each tip in turn
        - ready to send to FITCH or equivalent C routine
}
\description{
Speficies how characters are stored in a data object
}
\section{Functions}{
\itemize{
\item \code{TipsAreRows}: use if each row in a matrix corresponds to a tip

\item \code{TipsAreColumns}: use if each column in a matrix corresponds to a tip
}}

\keyword{internal}
