% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeDist-package.R
\docType{package}
\name{TreeDist-package}
\alias{TreeDist}
\alias{TreeDist-package}
\title{TreeDist: Distances between Phylogenetic Trees}
\description{
Implements measures of tree similarity, including 
  Information-based generalized Robinson-Foulds distances
  (Mutual Phylogenetic Information, Mutual Clustering Information; Smith, in revision); 
  the metric of Nye et al. (2006) <doi:10.1093/bioinformatics/bti720>;
  the Matching Split Distance (Bogdanowicz & Giaro 2012) <doi:10.1109/TCBB.2011.48>;
  the Kendall-Colijn (2016) distance <doi:10.1093/molbev/msw124>, and the
  Nearest Neighbour Interchange (NNI) distance, approximated per Li et al. 1996
  <doi:10.1007/3-540-61332-3_168>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeDist}
  \item \url{https://github.com/ms609/TreeDist}
  \item Report bugs at \url{https://github.com/ms609/TreeDist/issues}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (0000-0001-5660-1727) [copyright holder, programmer]

Other contributors:
\itemize{
  \item Roy Jonker \email{roy_jonker@magiclogic.com} [programmer]
  \item Harold Cooper \email{hbc@mit.edu} [contributor]
  \item Yong Yang \email{yongyanglink@gmail.com} [contributor]
}

}
\keyword{internal}
