% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_info.R, R/tree_distance_mmsi.R
\name{TreeDistance}
\alias{TreeDistance}
\alias{SharedPhylogeneticInfo}
\alias{DifferentPhylogeneticInfo}
\alias{ClusteringInfoDistance}
\alias{ClusteringInfoDist}
\alias{ExpectedVariation}
\alias{MutualClusteringInfo}
\alias{MutualClusteringInformation}
\alias{SharedPhylogeneticInfoSplits}
\alias{MutualClusteringInfoSplits}
\alias{MatchingSplitInfo}
\alias{MatchingSplitInfoDistance}
\alias{MatchingSplitInfoSplits}
\title{Information-based generalized Robinson-Foulds distances}
\usage{
TreeDistance(tree1, tree2 = tree1)

SharedPhylogeneticInfo(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

DifferentPhylogeneticInfo(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

ClusteringInfoDistance(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

ExpectedVariation(tree1, tree2, samples = 10000)

MutualClusteringInfo(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

SharedPhylogeneticInfoSplits(splits1, splits2, nTip = attr(splits1,
  "nTip"), reportMatching = FALSE)

MutualClusteringInfoSplits(splits1, splits2, nTip = attr(splits1,
  "nTip"), reportMatching = FALSE)

MatchingSplitInfo(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

MatchingSplitInfoDistance(tree1, tree2 = tree1, normalize = FALSE,
  reportMatching = FALSE)

MatchingSplitInfoSplits(splits1, splits2, nTip = attr(splits1, "nTip"),
  reportMatching = FALSE)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, with leaves labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{samples}{Integer specifying how many samplings to obtain;
accuracy of estimate increases with \code{sqrt(samples)}.}

\item{splits1, splits2}{Logical matrices where each row corresponds to a leaf,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a split, such that each leaf is identified as
a member of the ingroup (\code{TRUE}) or outgroup (\code{FALSE}) of the respective
split.}

\item{nTip}{Integer specifying the number of leaves in each split.}
}
\value{
If \code{reportMatching = FALSE}, the functions return a numeric
vector specifying the requested similarities or differences.

If \code{reportMatching = TRUE}, the functions additionally return details
of which clades are matched in the optimal matching, which can be viewed
using \code{\link{VisualizeMatching}}.
}
\description{
Tree similarity and distance measures that measure the amount of
phylogenetic or clustering information that two trees hold in common.
}
\details{
\href{https://ms609.github.io/TreeDist/articles/Robinson-Foulds.html#generalized-robinson-foulds-distances}{Generalized Robinson-Foulds distances}.
calculate tree similarity by finding an
optimal matching that the similarity between a split on one tree
and its pair on a second, considering all possible ways to pair splits
between trees (including leaving a split unpaired).

The methods implemented here use the concepts of
\href{https://ms609.github.io/TreeDist/articles/information.html}{entropy and information}
(MacKay 2003) to assign a similarity score between each pair of splits.

The returned tree similarity measures state the amount of information,
in bits, that the splits in two trees hold in common
when they are optimally matched, following Smith (forthcoming).
The complementary tree distance measures state how much information is
different in the splits of two trees, under an optimal matching.
}
\section{Functions}{
\itemize{
\item \code{SharedPhylogeneticInfo}: Shared phylogenetic information between splits of
two trees.

\item \code{DifferentPhylogeneticInfo}: Phylogenetic information difference between splits
of two trees.

\item \code{ClusteringInfoDistance}: Variation of clustering information distance between
splits of two trees.

\item \code{ExpectedVariation}: Estimate expected Shared and Different
Phylogenetic Information for a pair of trees of a given topology.

\item \code{MutualClusteringInfo}: Mutual Clustering Information of two trees.

\item \code{MatchingSplitInfo}: Mutual Matching Split information of two trees.

\item \code{MatchingSplitInfoDistance}: Variation of matching split information between two
trees.
}}

\section{Concepts of information}{


The phylogenetic (Shannon) information content and entropy of a split are
defined in
\href{https://ms609.github.io/TreeDist/articles/information.html}{a separate vignette}.

Using the mutual (clustering) information (Meila 2007, Vinh \emph{et al.} 2010) of
two splits to quantify their similarity gives rise to the Mutual Clustering
Information measure (\code{MutualClusteringInfo}); the entropy distance
gives the Clustering Distance \code{ClusteringInfoDistance}).
This approach is optimal in many regards, and is implemented, normalized
against the total information present, in the convenience function \code{TreeDistance}.

Using the amount of phylogenetic information common to two splits to measure
their similarity gives rise to the Shared Phylogenetic Information similarity
measure (\code{SharedPhylogeneticInfo}).  The amount of information distinct to
each of a pair of splits provides the complementary Different Phylogenetic
Information distance metric (\code{DifferentPhylogeneticInfo}).

The \code{MatchingSplitInfo} measure defines the similarity between a pair of
splits as the phylogenetic information content of the most informative
split that is consistent with both input splits; \code{MatchingSplitInfoDistance}
is the corresponding measure of tree difference.
\href{https://ms609.github.io/TreeDist/articles/Generalized-RF.html}{(More information here.)}
}

\section{Normalization}{


If \code{normalize = TRUE}, then results will be rescaled from zero to a nominal
maximum value, calculated thus:
\itemize{
\item \code{MutualClusteringInfo}, \code{ClusteringInfoDistance}: The sum of the entropy of
each split in each of the two trees.  See Vinh \emph{et al.} (2010, table 3) for
alternative normalization variants.
\item \code{SharedPhylogeneticInfo}, \code{MatchingSplitInfo}:
The sum of the information content of all splits in the least informative
tree. To scale against the information content of all splits in the most
informative tree, use \code{normalize = pmax}.
\item \code{DifferentPhylogeneticInfo}, \code{MatchingSplitInfoDistance}: The sum of the
phylogenetic information content of each split in each of the two trees.
}
}

\examples{
tree1 <- ape::read.tree(text='((((a, b), c), d), (e, (f, (g, h))));')
tree2 <- ape::read.tree(text='(((a, b), (c, d)), ((e, f), (g, h)));')
tree3 <- ape::read.tree(text='((((h, b), c), d), (e, (f, (g, a))));')

# Best possible score is obtained by matching a tree with itself
DifferentPhylogeneticInfo(tree1, tree1) # 0, by definition
SharedPhylogeneticInfo(tree1, tree1)
SplitwiseInfo(tree1) # Maximum shared phylogenetic information

# Best possible score is a function of tree shape; the splits within
# balanced trees are more independent and thus contain less information
SplitwiseInfo(tree2)

# How similar are two trees?
SharedPhylogeneticInfo(tree1, tree2) # Amount of phylogenetic information in common
VisualizeMatching(SharedPhylogeneticInfo, tree1, tree2) # Which clades are matched?

DifferentPhylogeneticInfo(tree1, tree2) # Distance measure
DifferentPhylogeneticInfo(tree2, tree1) # The metric is symmetric

# Are they more similar than two trees of this shape would be by chance?
ExpectedVariation(tree1, tree2, sample=12)['DifferentPhylogeneticInfo', 'Estimate']

# Every split in tree1 is contradicted by every split in tree3
# Non-arboreal matches contain clustering, but not phylogenetic, information
SharedPhylogeneticInfo(tree1, tree3) # = 0
MutualClusteringInfo(tree1, tree3) # > 0


}
\references{
\itemize{
\item \insertRef{Mackay2003}{TreeDist}
\item \insertRef{Meila2007}{TreeDist}
\item \insertRef{SmithDist}{TreeDist}
\item \insertRef{Vinh2010}{TreeDist}
}
}
\seealso{
Other tree distances: \code{\link{JaccardRobinsonFoulds}},
  \code{\link{KendallColijn}}, \code{\link{MASTSize}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NNIDist}}, \code{\link{NyeTreeSimilarity}},
  \code{\link{RobinsonFouldsInfo}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith} (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree distances}
