% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance.R, R/tree_distance_mmsi.R
\name{TreeDistance}
\alias{TreeDistance}
\alias{MutualPhylogeneticInfo}
\alias{VariationOfPhylogeneticInfo}
\alias{VariationOfClusteringInfo}
\alias{ExpectedVariation}
\alias{MutualClusteringInfo}
\alias{MutualPhylogeneticInfoSplits}
\alias{MutualMatchingSplitInfoSplits}
\alias{MutualClusteringInfoSplits}
\alias{MutualMatchingSplitInfo}
\alias{VariationOfMatchingSplitInfo}
\title{Information-based generalized Robinson-Foulds distances}
\usage{
TreeDistance(tree1, tree2)

MutualPhylogeneticInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

VariationOfPhylogeneticInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

VariationOfClusteringInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

ExpectedVariation(tree1, tree2, samples = 1000)

MutualClusteringInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

MutualPhylogeneticInfoSplits(splits1, splits2, normalize = TRUE,
  reportMatching = FALSE)

MutualMatchingSplitInfoSplits(splits1, splits2, reportMatching = FALSE)

MutualClusteringInfoSplits(splits1, splits2, normalize = TRUE,
  reportMatching = FALSE)

MutualMatchingSplitInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)

VariationOfMatchingSplitInfo(tree1, tree2, normalize = FALSE,
  reportMatching = FALSE)
}
\arguments{
\item{tree1, tree2}{Trees of class \code{phylo}, with tips labelled identically,
or lists of such trees to undergo pairwise comparison.}

\item{normalize}{If a numeric value is provided, this will be used as a
maximum value against which to rescale results.
If \code{TRUE}, results will be rescaled against a maximum value calculated from
the specified tree sizes and topology, as specified in the 'Normalization'
section below.
If \code{FALSE}, results will not be rescaled.}

\item{reportMatching}{Logical specifying whether to return the clade
matchings as an attribute of the score.}

\item{samples}{Integer specifying how many samplings to obtain;
accuracy of estimate increases with \code{sqrt(samples)}.}

\item{splits1, splits2}{Logical matrices where each row corresponds to a terminal,
either listed in the same order or bearing identical names (in any sequence),
and each column corresponds to a bipartition split, such that
each terminal is identified as a member of the ingroup (\code{TRUE}) or outgroup
(\code{FALSE}) of the respective bipartition split.}
}
\value{
If \code{reportMatching = FALSE}, the functions return a numeric
vector specifying the requested similarities or differences.

If \code{reportMatching = TRUE}, the functions additionally return details
of which clades are matched in the optimal matching, which can be viewed
using \code{\link{VisualizeMatching}}.
}
\description{
Tree similarity and distance measures that measure the amount of
phylogenetic or clustering information that two trees hold in common.
}
\details{
A tree similarity measure is generated by finding an optimal matching that
maximises the total information in common between a partition on one tree
and its pair on a second, considering all possible ways to pair partitions
between trees (including leaving a partition unpaired).

The returned tree similarity measures state the amount of information,
in bits, that the partitions in two trees hold in common
when they are optimally matched, following Smith (forthcoming).
The complementary tree distance measures state how much information is
different in the partitions of two trees, under an optimal matching.
}
\section{Functions}{
\itemize{
\item \code{MutualPhylogeneticInfo}: Mutual phylogenetic information between two trees.

\item \code{VariationOfPhylogeneticInfo}: Variation of phylogenetic information between two trees.

\item \code{VariationOfClusteringInfo}: Variation of clustering information between two trees.

\item \code{ExpectedVariation}: Estimate expected Variation of
Phylogenetic Information and Mutual Phylogenetic Information for a pair of trees of
a given topology.

\item \code{MutualClusteringInfo}: Mutual Clustering Information of two trees.

\item \code{MutualPhylogeneticInfoSplits}: Calculate mutual phylogenetic information from splits instead of trees.

\item \code{MutualMatchingSplitInfoSplits}: Calculate variation of matching split information from splits instead of trees.

\item \code{MutualClusteringInfoSplits}: Calculate clustering information from splits instead of trees

\item \code{MutualMatchingSplitInfo}: Mutual Matching Split information of two trees.

\item \code{VariationOfMatchingSplitInfo}: Variation of matching split information between two trees.
}}

\section{Concepts of information}{


Each partition in a tree can be viewed as a statement that the 'true'
tree is one of those that splits the taxa as specified.  This concept
corresponds to the concept of phylogenetic information, which views the
information content of a pair of partitions as relating to the proportion of
phylogenetic trees that are consistent with both partitions.

This proportion gives rise to the Mutual Phylogenetic Information similarity
measure (\code{MutualPhylogeneticInfo}), and the complementary
Variation of Phylogenetic Information distance metric
(\code{VariationOfPhylogeneticInfo}).

An information-theoretic take on the \code{\link{MatchingSplitDistance}} offers a
relaxed interpretation of this measurement, which attributes some information
to similar yet contradictory splits.  This metric is implemented in
\code{MutualMatchingSplitInfo} and \code{VariationOfMatchingSplitInfo}.

Alternatively, each partition can be viewed as a statement that the taxa
are subdivided into the two groups specified.  On this view, the
information content of a pair of partitions corresponds to their
mutual clustering information (Meila 2007, Vinh2010), giving rise to their
Mutucal Clustering Information and Variation of Clustering Information
similarity/distance metrics (\code{MutualClusteringInfo}, \code{VariationOfClusteringInfo}).

This latter approach is optimal in many regards, and is implemented, normalized
against the total information present, in the convenience function \code{TreeDistance}.
}

\section{Normalization}{


If \code{normalize = TRUE}, then results will be rescaled from zero to a nominal
maximum value, calculated thus:
\itemize{
\item \code{MutualPhylogeneticInfo}, \code{MutualClusteringInfo}, \code{MutualMatchingSplitInfo}:
The information content of the least informative tree.
To scale against the information content of the most informative tree, use
\code{normalize = pmax}.
\item \code{VariationOfPhylogeneticInfo}, \code{VariationOfClusteringInfo},
\code{VariationOfMatchingSplitInfo}: The sum of the
(phylogenetic or clustering) information content of the two trees.
}
}

\examples{
{
  tree1 <- ape::read.tree(text='((((a, b), c), d), (e, (f, (g, h))));')
  tree2 <- ape::read.tree(text='(((a, b), (c, d)), ((e, f), (g, h)));')
  tree3 <- ape::read.tree(text='((((h, b), c), d), (e, (f, (g, a))));')
  
  # Best possible score is obtained by matching a tree with itself
  VariationOfPhylogeneticInfo(tree1, tree1) # 0, by definition
  MutualPhylogeneticInfo(tree1, tree1)
  PartitionInfo(tree1) # Maximum mutual phylogenetic information
  
  # Best possible score is a function of tree shape; the partitions within
  # balanced trees are more independent and thus contain less information
  PartitionInfo(tree2)
  
  # How similar are two trees?
  MutualPhylogeneticInfo(tree1, tree2) # Amount of phylogenetic information in common
  VisualizeMatching(MutualPhylogeneticInfo, tree1, tree2) # Which clades are matched?
  VariationOfPhylogeneticInfo(tree1, tree2) # Distance measure
  VariationOfPhylogeneticInfo(tree2, tree1) # The metric is symmetric
  #'   
  # Are they more similar than two trees of this shape would be by chance?
  ExpectedVariation(tree1, tree2, sample=12)['VariationOfPhylogeneticInfo', 'Estimate']
  
  # Every partition in tree1 is contradicted by every partition in tree3
  # Non-arboreal matches contain clustering, but not phylogenetic, information
  MutualPhylogeneticInfo(tree1, tree3) # = 0
  MutualClusteringInfo(tree1, tree3) # > 0
  
}


}
\references{
{
\itemize{
\item \insertRef{Meila2007}{TreeDist}
\item \insertRef{SmithDist}{TreeDist}
\item \insertRef{Vinh2010}{TreeDist}
}
}
}
\seealso{
Other tree distances: \code{\link{KendallColijn}},
  \code{\link{MatchingSplitDistance}},
  \code{\link{NyeTreeSimilarity}}
}
\author{
Martin R. Smith
}
\concept{tree distances}
