% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSearch.R
\name{TreeSearch}
\alias{TreeSearch}
\title{Search for most parsimonious trees}
\usage{
TreeSearch(tree, dataset, method = "NNI", maxIter = 100, maxHits = 20,
  forestSize = 1, cluster = NULL, verbosity = 1, ...)
}
\arguments{
\item{tree}{a fully-resolved starting tree in \code{\link{phylo}} format, with the desired outgroup; edge lengths are not supported and will be deleted;}

\item{dataset}{A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}}, whose names 
correspond to the labels of any accompanying tree}

\item{method}{rearrangements to perform; one of \kbd{NNI}, \kbd{SPR}, or \kbd{TBR};}

\item{maxIter}{the maximum number of iterations to perform before abandoning the search;}

\item{maxHits}{the maximum times to hit the best pscore before abandoning the search;}

\item{forestSize}{the maximum number of trees to return - useful in concert with \code{\link{consensus}};}

\item{cluster}{a cluster prepared using \code{\link{PrepareCluster}}; may speed up search on multicore machines;}

\item{verbosity}{higher values provide more verbose user feedback in stdout;}

\item{\dots}{other arguments to pass to subsequent functions.}

\item{outgroup}{a vector listing the taxa in the outgroup;}

\item{concavity}{concavity constant for implied weighting (not currently implemented!);}
}
\value{
{
This function returns a tree, with an attribute \code{pscore} conveying its parsimony score.
Note that the parsimony score will be inherited from the tree's attributes, which is only valid if it 
was generated using the same \code{data} that is passed here.
}
}
\description{
Run standard search algorithms (\acronym{NNI}, \acronym{SPR} or \acronym{TBR}) 
to search for a more parsimonious tree.
}
\examples{
data('SigSut')
outgroup <- c('Lingula', 'Mickwitzia', 'Neocrania')
njtree <- Root(nj(dist.hamming(SigSut.phy)), outgroup, resolve.root=TRUE)
njtree$edge.length <- NULL; njtree<-SetOutgroup(njtree, outgroup)

\dontrun{
TreeSearch(njtree, SigSut.phy, outgroup, maxIter=20, method='NNI')
TreeSearch(njtree, SigSut.phy, outgroup, maxIter=20, method='SPR')
TreeSearch(njtree, SigSut.phy, outgroup, maxIter=20, method='TBR')}

}
\seealso{
\itemize{
\item \code{\link{Fitch}}, calculates parsimony score;
\item \code{\link{RootedNNI}}, conducts tree rearrangements;
\item \code{\link{SectorialSearch}}, alternative heuristic, useful for larger trees;
\item \code{\link{Ratchet}}, alternative heuristic, useful to escape local optima.
}
}
\author{
Martin R. Smith
}
\keyword{tree}
