% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeMatching}
\alias{VisualizeMatching}
\title{Visualise a matching}
\usage{
VisualizeMatching(Func, tree1, tree2, setPar = TRUE, precision = 3,
  Plot = plot.phylo, matchOptional = FALSE, ...)
}
\arguments{
\item{Func}{Function used to construct tree similarity.}

\item{tree1, tree2}{Trees of class \code{phylo}, with tips labelled identically.}

\item{setPar}{Logical specifying whether graphical parameters should be
set to display trees side by side.}

\item{precision}{Integer specifying number of significant figures to display
when reporting matching scores.}

\item{Plot}{Function to use to plot trees.}

\item{matchOptional}{Logical specifying whether to pair partitions with zero
similarity (\code{TRUE}), or leave them unpaired (\code{FALSE}).}

\item{\dots}{Additional parameters to send to \code{Plot}.}
}
\description{
Depicts the bipartitions that are matched between two trees using a
specified Generalized Robinson-Foulds similarity measure.
}
\details{
Note that when visualizing a Robinson-Foulds distance (using
\code{Func = RobinsonFoulds}), matched splits are assigned a \emph{similarity} score
of 1, which is deducted from the total number of splits to calculate the
Robinson-Foulds \emph{distance}.  Unmatched splits thus contribute one to total
tree distance.
}
\author{
Martin R. Smith
}
