% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_distance_utilities.R
\name{VisualizeMatching}
\alias{VisualizeMatching}
\title{Visualise a matching}
\usage{
VisualizeMatching(Func, tree1, tree2, setPar = TRUE, precision = 3,
  Plot = plot.phylo, ...)
}
\arguments{
\item{Func}{Function used to construct tree similarity.}

\item{tree1, tree2}{Trees of class \code{phylo}, with tips labelled identically.}

\item{setPar}{Logical specifying whether graphical parameters should be
set to display trees side by side.}

\item{precision}{Integer specifying number of significant figures to display
when reporting matching scores.}

\item{Plot}{Function to use to plot trees.}

\item{\dots}{Additional parameters to send to \code{Plot}.}
}
\description{
Depicts the bipartitions that are matched between two trees using a
specified Generalized Robinson Foulds tree distance measure.
}
\author{
Martin R. Smith
}
