% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VisualizeMatching}
\alias{VisualizeMatching}
\title{Visualise a matching}
\usage{
VisualizeMatching(Func, tree1, tree2, setPar = TRUE, precision = 3L,
  Plot = plot.phylo, matchZeros = TRUE, plainEdges = FALSE,
  edge.width = NULL, edge.color = NULL, ...)
}
\arguments{
\item{Func}{Function used to construct tree similarity.}

\item{tree1, tree2}{Trees of class \code{phylo}, with tips labelled identically.}

\item{setPar}{Logical specifying whether graphical parameters should be
set to display trees side by side.}

\item{precision}{Integer specifying number of significant figures to display
when reporting matching scores.}

\item{Plot}{Function to use to plot trees.}

\item{matchZeros}{Logical specifying whether to pair partitions with zero
similarity (\code{TRUE}), or leave them unpaired (\code{FALSE}).}

\item{plainEdges}{Logical specifying whether to plot edges with a uniform
width and colour (\code{TRUE}), or whether to draw edge widths according to the
similarity of the associated splits (\code{FALSE}).}

\item{edge.width, edge.color, \dots}{Additional parameters to send to \code{Plot}.}
}
\description{
Depicts the bipartitions that are matched between two trees using a
specified Generalized Robinson-Foulds similarity measure.
}
\details{
Note that when visualizing a Robinson-Foulds distance (using
\code{Func = RobinsonFoulds}), matched splits are assigned a \emph{similarity} score
of 1, which is deducted from the total number of splits to calculate the
Robinson-Foulds \emph{distance}.  Unmatched splits thus contribute one to total
tree distance.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
